### 2021-10-07 #470 @ewallace
Submission tidyqpcr - Quantitative PCR analysis in the tidyverse
Submitting Author Name: Edward Wallace
Submitting Author Github Handle: @ewallace
Other Package Authors Github handles: @DimmestP
Repository:  https://github.com/ewallace/tidyqpcr
Version submitted: 0.3.0
Submission type: Standard
Reviewers: @kelshmo
Version accepted: TBD

---

```
Package: tidyqpcr
Title: Quantitative PCR Analysis with the Tidyverse
Version: 0.3
Authors@R: c(person("Edward", "Wallace", email = "Edward.Wallace@ed.ac.uk", role = c("aut", "cre")),
             person("Sam", "Haynes", email = "samuel.haynes10@gmail.com", role = c("ctb")))
Description: This package is for reproducible quantitative PCR (qPCR) analysis using packages from the tidyverse, notably dplyr and ggplot2. It normalizes (by ddCq), summarizes, and plots pre-calculated Cq data, and plots raw amplification and melt curves from Roche LightCycler machines. It does NOT (yet) calculate Cq data from amplification curves.
    R (>= 3.1.0),
    tidyr
    forcats,
    assertthat,
    lifecycle
Suggests: knitr,
    tidyverse,
    cowplot,
License: Apache License 2.0 + file LICENSE
Language: en-GB
```

tidyqpcr is an R package that empowers scientists to conduct reproducible, flexible, and best-practice compliant quantitative polymerase chain reaction (qPCR) analysis.

tidyqpcr offers a standardised user interface and structure for qPCR analysis, within the tidyverse paradigm of spreadsheet-like rectangular data frames and generic functions that build up complex analyses in a series of simple steps.

Any molecular biologist or bioinformatician who needs to design or analyse a qPCR experiment.

Quantitative PCR is among the most common techniques in biological and biomedical re-
search, used for the quantification of DNA and RNA.

Standardised and open-source qPCR analysis pipelines will encourage best-practices in the reporting of qPCR results, improve the evaluation of qPCR experiments and ultimately lead to increased confidence in conclusions based on qPCR data.

Some open-source libraries for qPCR analysis are available, notably [qpcR](https://github.com/anspiess/qpcR) (Spiess, 2018) and [pcr](https://github.com/MahShaaban/pcr) (Ahmed & Kim, 2018). qpcR is a feature rich qPCR analysis package relying on an object-oriented approach using S4 classes. pcr is a less extensive qPCR analysis package based on the tidyverse suite of generic data-science tools using the paradigm of tidy data (spreadsheet-like rectangular data frames). However, available packages either assume extensive prior R knowledge, overlook best-practices in qPCR experiments, or lack extensive documentation. There remains a need for a qPCR analysis package that integrates with the user-friendly tidyverse, encourages the use of MIQE best-practice compliant experimental design, and provides detailed example analysis pipelines as R vignettes.

We presented tidyqpcr on the rOpenSci discussion page which [Stefanie Butland and Sean Hughes](https://discuss.ropensci.org/t/tidyqpcr-for-quantitative-pcr-analysis-pre-presubmission-community-enquiry/2056) kindly responded to. In response to @seaaan ’s comments we improved the vignettes, stuck to a consistent function naming convention and added functionality to calculate ddCq. We intend to add further functionality in future versions including: support for absolute quantification, support for multiple targets per well, and enabling the use of the plater package.

In email correspondence with Stefanie, we believe Julia Gustavsen would be a perfect editor for our project as they reviewed Sean’s [plater](https://github.com/ropensci/software-review/issues/60) package. As for reviewers, we think someone with experience in conducting assays for RNA/DNA quantification and normalisation would be of benefit because of the emphasis on experimental design best-practices.

- [ ] Do you wish to submit an Applications Article about your package to [Methods in Ecology and Evolution](http://besjournals.onlinelibrary.wiley.com/hub/journal/10.1111/(ISSN)2041-210X/)?

- [x] Do you wish to submit an Applications Article about your package to Journal of Open Source Software?

We have included a paper.md file in the repository as per JOSS instructions to authors.
