### 2021-10-08 a3da7a6
Adds R-CMD-check CI
