### 2021-10-08 667a7af
Add foward and reverse R version check
