### 2021-10-08 336b7cc
Remove duplicate badges
