### 2021-10-08 43b1473
Updates contribution section in README
