### 2021-10-13 f19fd6c
Adds examples to every tidyqpcr function
