+++ man/create_rowkey_4_in_16.Rd
@@ -16,7 +16,7 @@ tibble (data frame) with 16 rows, and variables well_row, prep_type,
 }
 \description{
 Create a 16-row key with 4 values repeated over 16 plate rows. Each of the 4
-values is repeated over 3x +RT technical replicates and 1x -RT.
+values is repeated over 3x +RT Techreps and 1x -RT.
 }
 \examples{
 create_rowkey_4_in_16(sample_id=c("sheep","goat","cow","chicken"))
