### 2021-11-20 #470 @kelshmo
---

The name of the package is clear and I appreciate the thorough descriptive content in your README.

#### Notes on README
 - I suggest defining MIQE
 - There is a broken link to Rtools [here](https://github.com/ewallace/tidyqpcr#for-windows-users)
 - It looks like you have a github pages setup. I have seen pages linked to in the "about" repository section. I couldn't find it linked in the README.

####Vignettes
"Primers and probes calibration vignette" and "Multifactorial multi-plate qPCR analysis example"
- Vignette needs `ggplot2` loaded to run. For clarity, I might call `library(cowplot)` in the first code chunk too.
- `/inst` will not exist for someone who has installed the package. There are a few references to files in `/inst`. Instead, you can use `system.file()` to find this file like you did at the beginning of the vignette: `read_lightcycler_1colour_cq(
  "../inst/extdata/Edward_qPCR_Nrd1_calibration_2019-02-02_Ct.txt"
)`

No comments on the other vignettes!

#### Notes on contributing
- I like the use of "good first issue"!

#### Notes on function definitions
- Five functions have the same title "Create a blank plate template as a tibble". I can infer from the function name what is distinct, but it would be helpful to update the titles to be unique. Same for five functions titled "Nice tick labels for logarithmic axes in ggplot2."
- `label_plate_rowcol` refers to a vignette but is not explicit which one is relevant.

#### Notes on examples
- The example for `calculate_efficiency` returned the warning `Warning message: Unknown or uninitialised column: `target_id`. `
- I was confused when the documentation for `create_blank_plate()` populated when I was looking for `make_row_names_echo1536()`. This is definitely canonical but it could be helpful in the description to add some context (1 or 2 sentences).

Thank you Edward and Sam, nice work and great package! I'm happy to learn about codemeta.json too. I appreciate your patience with my review.
