### 2021-11-24 #470 @wolski
@DimmestP
Dear package developers. I fully agree with the comments made by @kelshmo. However, if they are a relatively quick fix, then please fix them quickly!  First, we could avoid redundant comments from my side, and I could more easily focus on different aspects of your package.

For instance, I typically go through vignettes by executing code chunks, and relative paths don't work.

```
> plate_cq_extdata <- read_lightcycler_1colour_cq(
+   "../inst/extdata/Stuart_dAgr_glyS_spoVG_5S_individualWells_Cq.txt"
+   ) %>%
+   dplyr::filter(! well %in% wells_exclude ) %>% # exclude blank wells
+   dplyr::select(well, sample_info, cq)
 Error: '../inst/extdata/Stuart_dAgr_glyS_spoVG_5S_individualWells_Cq.txt' does not exist in current working directory ('C:/Users/wolski/prog/tidyqpcr').
```
