### 2021-11-30 ewallace/tidyqpcr#121 @ewallace
Clarify README from @kelshmo review
Responding to @kelshmo's review, https://github.com/ropensci/software-review/issues/470#issuecomment-974548665

In README.md
* [x] Define MIQE
* [x] Fix [broken link to Rtools](https://github.com/ewallace/tidyqpcr#for-windows-users)
* [ ] Link to GitHub pages setup.
