### 2021-11-30 ewallace/tidyqpcr#122 @ewallace
Vignette packages and files for @kelshmo review
Responding to @kelshmo's review, https://github.com/ropensci/software-review/issues/470#issuecomment-974548665

In "Primers and probes calibration vignette" and "Multifactorial multi-plate qPCR analysis example"

* [ ] Vignette needs ggplot2 loaded to run. For clarity, I might call library(cowplot) in the first code chunk too.
* [ ] `/inst` will not exist for someone who has installed the package. There are a few references to files in `/inst`. Instead, you can use `system.file()` to find this file like you did at the beginning of the vignette: `read_lightcycler_1colour_cq( "../inst/extdata/Edward_qPCR_Nrd1_calibration_2019-02-02_Ct.txt" )`

@wolski also raised the point about file paths, so that is a priority. See https://github.com/ropensci/software-review/issues/470#issuecomment-978172879
