### 2021-11-30 ewallace/tidyqpcr#124 @ewallace
Function documentation from @kelshmo review
Responding to @kelshmo's review, https://github.com/ropensci/software-review/issues/470#issuecomment-974548665

Notes on function definitions and examples

* [ ] Five functions have the same title "Create a blank plate template as a tibble". I can infer from the function name what is distinct, but it would be helpful to update the titles to be unique.
* [ ] Same for five functions titled "Nice tick labels for logarithmic axes in ggplot2." (EW: I don't know if this is important enough to fix? It's not core functionality.)
* [ ] `label_plate_rowcol` refers to a vignette but is not explicit which one is relevant.
* [ ] The example for `calculate_efficiency` returned the warning `Warning message: Unknown or uninitialised column: target_id`.
* [ ] I was confused when the documentation for `create_blank_plate()` populated when I was looking for `make_row_names_echo1536()`. This is definitely canonical but it could be helpful in the description to add some context (1 or 2 sentences).
