### 2021-12-03 #470 @wolski
@ewallace @DimmestP

I already have a few observations about what you have written about tidyqpcr and other packages in the section:

 "Are there other R packages that accomplish the same thing? If so, how does yours differ or meet our criteria for best-in-category?"

You explicitly mention the package qpcR and pcr.

Your write "qPCR analysis package relying on an object-oriented approach using S4 classes."
I am not a big fan of APIs with "@" either.  But are you sure you are talking about qpcR? I checked the CRAN page and observed no dependency on the "methods".  I downloaded the tar.gz and searched the sources for the "setClass" and "@" string, and there are no hits. "qpcR" is not written in S4.

Furthermore, why do you say that "available packages either assume extensive prior R knowledge"?
Does this indicate a GUI for `tidyqpcr`; a shiny application?  Or can I accomplish the same task with tidyqpcr with fewer lines of code than when using the PCR package? Finally, the vignettes you include are not concise; on the contrary, you heavily use ggplot2 in the vignette, while ggplot2 does not have a flat learning curve.

I do not know if @seaaan means the same as I when he writing "Focus on functions provided by your package and what they do and how to use them" (e.g. in the multifactorial vignette, explain at greater length about normalizeqPCR, debaseline, getRdTall)", but I would say the same.
https://discuss.ropensci.org/t/tidyqpcr-for-quantitative-pcr-analysis-pre-presubmission-community-enquiry/2056
This comment was made in May 2021 i.e., half a year ago, but I do not see any evidence of how you addressed them in the vignettes of tidyqpcr.

Also, I would like to understand better what do you mean "overlook best-practices in qPCR experiments"? Indeed, you refer to the "MIQE best-practice compliant experimental design" which the alternatives can't meet. MIQE is a checklist of requirements to publish a qPCR experiment.
There is Table 1 in the publication about MIQE you have referenced. Would you maybe tabulate which MIQE requirements "tidyqpcr" already covers using which functions, which would improve the package documentation? I am aware that this is a lot to ask, but could you please include the CRAN alternatives in this table to give substance to your claim that those alternatives of tidyqpcr overlook best practices. Most likely, you can reuse this table in an upcoming publication of `tidyqpcr`.

Finally, you say  "or lack extensive documentation".
Both of the packages you mention are, in my opinion, well documented, and both are published in peer-reviewed journals:
https://peerj.com/articles/4473/
https://academic.oup.com/bioinformatics/article/24/13/1549/238435
I am not claiming that this guarantees their utility but documented they are.

Furthermore, a web search for R and qPCR points to several more R packages that you do not mention:
https://bmcgenomics.biomedcentral.com/articles/10.1186/1471-2164-13-296
   https://www.bioconductor.org/packages/release/bioc/html/HTqPCR.html

I feel that you did not address the points in
https://devguide.ropensci.org/policies.html#overlap sufficiently, and I would welcome a critical comparison.

I will have a closer look at the three first points during the next week, but I have a few observations regarding:
 "Actively maintained while alternatives are poorly".

qPCR has a release history of 14 years; it has one reverse `depends` and two reverse `imports`, and four reverse `suggests`. On the other hand, the pcr package has a release history of 4 years and an active github repository.

Am I correct in assuming that `tidyqpcr` is a Ph.D. project of Samuel?
Samuel is a  gifted programmer, and for sure, you can publish this package in one of the many specialized journals. But how sustainable is this project? I am asking because `tidyqpcr` is by no means feature-complete (as the long list of To-Do's in README.md indicates).
I find it worrying that there has been practically no activity within this project since 2019
https://github.com/ewallace/tidyqpcr/graphs/code-frequency

Furthermore, since their first release, all three other pinned projects on github.com/ewallace have not been significantly developed.
Protocols were released in 2016 with about two commits/year since;
stochsimode has seen no commits for six years. RoboViz has not had a single commit for three years.
