### 2021-12-05 #470 @jooolia
Thanks for the questions @wolski. I will let the authors reply. One thing that I want to note is that currently we do not require any discussion of the lifecycle of a package like we do for our statistical  software review, but it is under discussion here: https://github.com/ropensci/dev_guide/issues/366
