### 2021-12-06 ewallace/tidyqpcr#125 @ewallace
Tabulate MIQE requirements addressed by tidyqpcr, from @wolski review
Software review by @wolski https://github.com/ropensci/software-review/issues/470#issuecomment-985536546
suggests:

> Also, I would like to understand better what do you mean "overlook best-practices in qPCR experiments"? Indeed, you refer to the "MIQE best-practice compliant experimental design". MIQE is a checklist of requirements to publish a qPCR experiment.
There is Table 1 in the publication about MIQE you have referenced. Would you maybe tabulate which MIQE requirements "tidyqpcr" already covers using which functions, which would improve the package documentation? I am aware that this is a lot to ask, but could you please include the CRAN alternatives in this table to give substance to your claim that those alternatives of tidyqpcr overlook best practices. Most likely, you can reuse this table in an upcoming publication of tidyqpcr.

Let's do that.
