### 2021-12-06 ewallace/tidyqpcr#126 @ewallace
Rewrite description of other packages from @wolski review
Software review by @wolski https://github.com/ropensci/software-review/issues/470#issuecomment-985536546
points out that we have described other packages inaccurately.

This needs substantial effort and rewrites to `paper.md`.

> "Are there other R packages that accomplish the same thing? If so, how does yours differ or meet our criteria for best-in-category?"

> You explicitly mention the package qpcR and pcr.

> Your write "qPCR analysis package relying on an object-oriented approach using S4 classes."
> I am not a big fan of APIs with "@" either. But are you sure you are talking about qpcR? I checked the CRAN page and observed no dependency on the "methods". I downloaded the tar.gz and searched the sources for the "setClass" and "@" string, and there are no hits. "qpcR" is not written in S4.

We need to correct this, they use S3 classes not S4 classes.

> Furthermore, why do you say that "available packages either assume extensive prior R knowledge"?
> Does this indicate a GUI for tidyqpcr; a shiny application? Or can I accomplish the same task with tidyqpcr with fewer lines of code than when using the PCR package? Finally, the vignettes you include are not concise; on the contrary, you heavily use ggplot2 in the vignette, while ggplot2 does not have a flat learning curve.

We should refocus away from other packages: tidyqpcr is really easy if you know tidyverse. That means users can do customisable analysis & visualisation in few lines of code. To some extent "tidyverse is a qPCR package for tidyverse users". We also aim to be a gateway to the utility of the tidyverse for qPCR users. Can we explain that better?

For me (@ewallace) I found the other packages we mention impossible because they didn't expose the data in a data frame that I understand. I am not a typical user, so we need to find a better way of saying that, that is generous to these other carefully-written packages.

> Finally, you say "or lack extensive documentation".
> Both of the packages you mention are, in my opinion, well documented, and both are published in peer-reviewed journals:
> https://peerj.com/articles/4473/
> https://academic.oup.com/bioinformatics/article/24/13/1549/238435
> I am not claiming that this guarantees their utility but documented they are.

We need to revisit this. Or if we are referring to _tutorials_ that tidyqpcr now has, we need to say something.

> Furthermore, a web search for R and qPCR points to several more R packages that you do not mention:
> https://bmcgenomics.biomedcentral.com/articles/10.1186/1471-2164-13-296
> https://www.bioconductor.org/packages/release/bioc/html/HTqPCR.html

> I feel that you did not address the points in
> https://devguide.ropensci.org/policies.html#overlap sufficiently, and I would welcome a critical comparison.

> I will have a closer look at the three first points during the next week, but I have a few observations regarding:
> "Actively maintained while alternatives are poorly".

> qPCR has a release history of 14 years; it has one reverse depends and two reverse imports, and four reverse suggests. On the other hand, the pcr package has a release history of 4 years and an active github repository.

In summary, we have to do a more thorough review of these other packages including their documentation, then address the overlap more thoroughly in the review thread.

My preference is to say no negative things about other packages, and focus on the positives of tidyqpcr.
