### 2021-12-06 ewallace/tidyqpcr#127 @ewallace
New vignette focusing on normalisation, delta_cq, etc? from @wolski review
Software review by @wolski https://github.com/ropensci/software-review/issues/470#issuecomment-985536546
Asked for addressing some key normalisation capabilities in the vignettes.

> I do not know if @seaaan means the same as I when he writing "Focus on functions provided by your package and what they do and how to use them" (e.g. in the multifactorial vignette, explain at greater length about normalizeqPCR, debaseline, getRdTall)", but I would say the same.
> https://discuss.ropensci.org/t/tidyqpcr-for-quantitative-pcr-analysis-pre-presubmission-community-enquiry/2056
> This comment was made in May 2021 i.e., half a year ago, but I do not see any evidence of how you addressed them in the vignettes of tidyqpcr.

Do we need a new vignette focusing on normalisation, delta_cq, etc.?
