### 2021-12-06 #470 @ewallace
@wolski Thank you for the thorough review.

You've made a lot of good points that our description of other packages was inaccurate, and we will address that at length. I created new issue ticket https://github.com/ewallace/tidyqpcr/issues/126, we will expand on that at some length and then get back to this thread.

Here I will answer quickly the points that can be answered quickly.

> Your write "qPCR analysis package relying on an object-oriented approach using S4 classes."
> I am not a big fan of APIs with "@" either. But are you sure you are talking about qpcR? I checked the CRAN page and observed no dependency on the "methods". I downloaded the tar.gz and searched the sources for the "setClass" and "@" string, and there are no hits. "qpcR" is not written in S4.

That's our error that we will correct, qpcR uses **S3** classes not S4 classes.

> I do not know if @seaaan means the same as I when he writing "Focus on functions provided by your package and what they do and how to use them" (e.g. in the multifactorial vignette, explain at greater length about normalizeqPCR, debaseline, getRdTall)", but I would say the same.
> https://discuss.ropensci.org/t/tidyqpcr-for-quantitative-pcr-analysis-pre-presubmission-community-enquiry/2056
> This comment was made in May 2021 i.e., half a year ago, but I do not see any evidence of how you addressed them in the vignettes of tidyqpcr.

That is fair, we have not expanded on these in the vignettes since May. What would "good enough" look like here for you? Are there particular points that need explaining from your perspective?

For context, the vignettes we have focus on the pain points that users reported in interviews; and on complete example data analyses. The big pain points were about laying out samples on plates and describing samples and genes accurately. Since we worked on those vignettes, users have reported it's much easier to use tidyqpcr, so that was a victory. But we have not written vignette-style documentation on the normalisation procedures and ct specifically.

We created a new issue to discuss adding to the vignettes: https://github.com/ewallace/tidyqpcr/issues/127

> Am I correct in assuming that tidyqpcr is a Ph.D. project of Samuel?
> Samuel is a gifted programmer, and for sure, you can publish this package in one of the many specialized journals. But how sustainable is this project? I am asking because tidyqpcr is by no means feature-complete (as the long list of To-Do's in README.md indicates).
> I find it worrying that there has been practically no activity within this project since 2019
> https://github.com/ewallace/tidyqpcr/graphs/code-frequency

tidyqpcr is a collaboration that I started and that has become part of Sam's PhD. We are planning to continue it after he graduates, and preparing for that by following best practices for software documentation and maintainability. Our next steps are:

- complete this review (quality)
- publish in ropensci and JOSS (visibility, helps with future funding)
- prepare a tutorial workshop including hopefully a video (builds user base)
- advertise for a project student to add features.

The recent low activity on tidyqpcr is due to Sam focusing on another manuscript with analysis that makes extensive use of tidyqpcr, indeed substantiates the claims that tidyqpcr can take data all the way to publishable figures:
https://github.com/DimmestP/chimera_project_manuscript

That manuscript is preprinted and will soon be submitted to a journal.

> Furthermore, since their first release, all three other pinned projects on github.com/ewallace have not been significantly developed.
> Protocols were released in 2016 with about two commits/year since;
> stochsimode has seen no commits for six years. RoboViz has not had a single commit for three years.

You've identified a problem with my pinned projects, not so much with the software dev.

Ironically or otherwise, [stochsimcode](https://github.com/ewallace/stochsimcode) was _my_ PhD code, which I put on GitHub because I was still getting occasional code requests from students doing masters projects.

riboviz has been under intensive development after moving to its own organisation [riboviz/riboviz](https://github.com/riboviz/riboviz).

So a question for @wolski: what evidence would convince you of the future of this software? Is it a development plan with timeline? Or something else?
