### 2021-12-10 #470 @wolski
# Best in category?

Some thoughts about the "best in category."
Making a judgment about the best in the category would require extensive formal testing of all the relevant packages to avoid biases.
Without it, the judgment would rely on heuristics only and be biased. Still, it is challenging to eliminate biases because of the complexity of package developments.
As soon as I have more time, I will try to discuss, and hopefully, better understand the ropensci review process at https://discuss.ropensci.org/.

Here is my attempt to formalize the review.

## Field and laboratory reproducibility tools
The package tidyqpcr was submitted in the category
field and laboratory reproducibility tools.:

The family of "creation functions" of _tidyqpcr_ falls into
"Automation of field and lab protocols, such as sample tracking and tagging,  form and datasheet generation,"
subcategory? My choice for this, however, would be PlateDesigner.
https://www.ncbi.nlm.nih.gov/pmc/articles/PMC6821189/
for ease of use.

The functions:
read_lightcycler_1colour_raw, ead_lightcycler_1colour_cq
are in the subcategory:
"interfacing with laboratory equipment or information systems, and executing experimental designs."
The package `ReadqPCR` has more functionality in this subcategory.

But there is also a group of functions with names starting with scale_ and ending with nice, which relate to data visualization.

Finally, a group of functions starts with caculate_ which is statistics? Again, the CRAN package pcr implements a more comprehensive set and would be my preferred choice.

# Package name

A comment regarding the package name tidy. _tidy_ isn't just about using tidy data tables. It is also about grouping functions into tidy packages, one for plotting, a different one for factors, time, etc. tidyqpcr is an analysis script that reads analyses, and visualizes data, converted into a package. It feels pretentious to include the prefix _tidy_ in the name of this package.

# Past reviews

I hope there is no requirement to rewrite the suggestion made by @seaaan, more than 6 months ago, in my own words, to include it into this review process.
Except for a few, those comments are still not addressed by the authors of tidyqpcr.  I am not saying that you should implement all the recommendations, but I would appreciate it if you could explain why you decided not to.

I am listing them here:

Feature suggestions:
- Determine Ct by setting a threshold on amp curve data
   - Functions to calculate dCt (I believe normalizeqPCR does this) and ddCt
   - Support for multiple targets per well
   - Functions for common plots (e.g. plot_amp_curve, plot_melt_curve), though I acknowledge that these may be tricky (figuring out what to facet by/color by/etc)
   - Support for absolute quantification via standard curves
- Vignettes:
   - Focus on functions provided by your package and what they do and how to use them (e.g. in the multifactorial vignette, explain at greater length about normalizeqPCR, debaseline, getRdTall)
   - Explain a bit more about the format the data needs to be in to be read into the package (since qPCR machines produce data in a variety of formats)
   - Try to simplify them a bit, they're a little overwhelming at the moment (for example, in the multifactorial experiment vignette, drop thiolutin [six conditions becomes four] and some of the primer sets, only show one plot of melt curves and one set of amp curves).

https://discuss.ropensci.org/t/tidyqpcr-for-quantitative-pcr-analysis-pre-presubmission-community-enquiry/2056

# My comments

## Vignettes
Rmarkdown and vignette title should be the same:
You have:
title: "Primers and probes calibration vignette"
 %\VignetteIndexEntry{PrimerCalibration}
but better
title: "Primers and probes calibration vignette"
 %\VignetteIndexEntry{Primers and probes calibration vignette}

The vignettes
"Delta Cq 96-well plate qPCR analysis example" and
"Introduction to designing an experiment and setting up a plate plan in tidyqpcr" is fine, good balance of code and text. However, there are almost no figure captions and text explaining the code in the vignettes:
"Primers and probes calibration vignette," "Multifactorial multi-plate qPCR analysis example."

However, I have the impression that there are many redundancies in the vignettes, e.g., the `display_plate` function and the experiment setup are repeatedly covered.
Simplifying the examples as @seaaan suggested will also help you meet the 5MB limit required by either CRAN or Bioconductor.

## Connecting plate design and data reading.

The package contains many functions to create a plate design. They are then used to fill the plates using a multichannel pipette.
The created annotations (e.g., WT, C+ ) can also be 'fed' into the measuring device. The sample_info column contains the annotation which was generated in the first step. Correct?

```
##    well  sample_info      cq
##    <chr> <chr>         <dbl>
##  1 A1    WT OD0.5 glyS  14.0
##  2 A2    WT OD0.5 glyS  13.9
##  3 A3    WT OD0.5 glyS  14.0
```

Why can't you then not reproduce the plate design from the sample_info? Why is this code needed?

```
plate_cq_data <-
  create_blank_plate_96well() %>% # row and column labels for each well
  inner_join(plate_cq_extdata) %>% # join that info to loaded data
  # next separate the sample and target information for future data analysis
  separate(sample_info,
           into = c("strain","OD","target_id"),
           sep = " ",
           remove = FALSE) %>%
  unite("sample_id",strain, OD, remove = FALSE) %>% # put sample_id together
  mutate(prep_type = "+RT") # add information on cDNA prep type
```

I would like to have code like:

```
plate1 <- read_lightcycler_1colour_cq
display_plate(plate1)
````

## display_plate
It always seems to color the wells according to the target_ID, and there is no option to choose a different column. Could you add an argument allowing to specify that column?

## Confounding

If I am correct, thermocyclers are used to perform qPCR experiments. Aren't the wells in the plates' center exposed to different temperatures than on the edges? If so, how this impacts the quantification results? Does your design function takes this into account, and if not, why? Is there a theoretical option to block for the confounding of plate position?

## Visualizing measurement results in the plate layout.

Would you please provide functions to visualize the results in plate layout, e.g., how those quantitation cycle, delta Cq looks on the plate? This would allow visualizing if there is a plate position effect.

# Conclusion "Jack of all trades and a master of none."

_tidyqpcr_ resembles an analysis script, with some helper function, then an R package. It tries to do many things but does not impress at any.

One possible direction for this package, would be to integrate several specialized packages, e.g.  PlateDesigner, ReadqPCR, CRAN:pcr, using a common data structure (can be a tidy table). This integration would make it easier to perform  MIQE compatible analysis. In addition, this package could implement functions to visualize the data easily. A possible implementation would start by annotating the tidy table, i.e., defining which columns contain plate positions, response variables, and explanatory variables. Then, using this information, figures (similar to those in the vignettes) could be created with a function call.

@jooolia, these two are experts in qPCR, which I am not. @seaaan, @MahShaaban
