### 2021-12-11 #470 @jooolia
Dear @wolski , thank you for the time that you have spent reviewing and the topics that you have raised for regarding this package and regarding the review process.

I would appreciate it if you could edit your latest comment so that it is more collegial. We are all here to help the authors improve their package by providing constructive comments. The sentences that you wrote in the Conclusion header and directly thereafter are not constructive.

Thank you for the experts in qpcr, however there would be a COI and thus they cannot be reviewers for this submission.

Finally, will you fill out the reviewer template (https://devguide.ropensci.org/reviewtemplate.html) or indicate that you are done with this round of your reviewing?

Thanks, Julia
