### 2021-12-13 #470 @wolski
@jooolia Thank you for your feedback, although I would have preferred to receive it by e-mail. But, yes, my comment "a script with a few helpers" without more context is not constructive. I was not referring to the package itself but the code equivalent. I am sorry.

I can not complete the review. I hypothesize that the reviewed package does duplicate large amounts of the functionality of other CRAN packages, which @ewallace must have known about in 2019, without making significant improvements. However, I do not have the time to prove or disprove it.

If open-source software does what I want, I will NOT start writing a new one. To me, there are those reasons to begin to write a new R package:
- The functionality is not available elsewhere in R in package form.
- To improve code quality and obtain a more robust and straightforward to maintain software.
- Provide a substantially more flexible API to enable new analysis or easier integration with other software components.
I think this is very much in line with:
"we aim to avoid duplication of functionality of existing R packages in any repo without significant improvements. An R package that replicates the functionality of an existing R package may be considered for inclusion in the rOpenSci suite if it significantly improves on alternatives in any repository (RO, CRAN, BioC) by being:".

Let's assume we were in 2019 when the development of the reviewed package started. The packages qPCR, pcr, plateR or plateDesigner existed and must have been known to @ewallace. Specifically, the `tidyqpcr::calculate_` functions are available elsewhere. Furthermore, their code isn't of low quality either, and it is debatable if tidyverse code, used in package implementations is easier to maintain.

Does the reviewed package provide a more flexible API or enable easier software components integration?  I hypothesize that it is possible to rewrite the vignettes of the package under review, using other packages available in 2019, and that the code might be even 'tidier' (allowing for a few helper and wrapper functions). That's why I said: "analysis script with a few helper functions," which represents the code equivalent of the reviewed package. Unfortunately, however, I do not have the time to prove or disprove it.
