### 2021-12-15 ewallace/tidyqpcr#128 @ewallace
Reduce package footprint below 5MB by compressing data, from @wolski comment
@wolski's review at https://github.com/ropensci/software-review/issues/470#issuecomment-991272373
says:

> Simplifying the examples as @seaaan suggested will also help you meet the 5MB limit required by either CRAN or Bioconductor.

We can almost certainly meet the size limit just by compressing the .txt files to .txt.gz files, and those can be called in exactly the same way. an easy fix.
