### 2021-12-15 e6284a6
gzipped inst/extdata, read .txt.gz in vignettes, Cq replaces Ct in filenames
