+++ inst/extdata/Stuart_dAgr_glyS_spoVG_5S_individualWells_Cq.txt
@@ -1,98 +1,0 @@
-Experiment: Stuart_WT_dAgr_spoVG_glyS_5S  Selected Filter: SYBR Green I / HRM Dye (483-533)
-Include	Color	Pos	Name	Cp	Concentration	Standard	Status
-True	255	A1	WT OD0.5 glyS	13.98		0	
-True	255	A2	WT OD0.5 glyS	13.94		0	
-True	255	A3	WT OD0.5 glyS	13.95		0	
-True	255	A4	WT OD1.5 glyS	14.55		0	
-True	255	A5	WT OD1.5 glyS	16.06		0	
-True	255	A6	WT OD1.5 glyS	14.44		0	
-True	255	A7	WT OD3 glyS	15.58		0	
-True	255	A8	WT OD3 glyS	15.55		0	
-True	255	A9	WT OD3 glyS	15.51		0	
-True	255	A10	WT OD6 glyS	12.82		0	
-True	255	A11	WT OD6 glyS	13.28		0	
-True	255	A12	WT OD6 glyS	13.25		0	
-True	255	B1	WT ODON glyS	19.33		0	
-True	255	B2	WT ODON glyS	19.00		0	
-True	255	B3	WT ODON glyS	18.91		0	
-True	255	B4	WT OD0.5 spoVG	13.58		0	
-True	255	B5	WT OD0.5 spoVG	13.59		0	
-True	255	B6	WT OD0.5 spoVG	13.59		0	
-True	255	B7	WT OD1.5 spoVG	12.82		0	
-True	255	B8	WT OD1.5 spoVG	12.79		0	
-True	255	B9	WT OD1.5 spoVG	12.91		0	
-True	255	B10	WT OD3 spoVG	13.67		0	
-True	255	B11	WT OD3 spoVG	13.70		0	
-True	255	B12	WT OD3 spoVG	13.66		0	
-True	255	C1	WT OD6 spoVG	12.27		0	
-True	255	C2	WT OD6 spoVG	11.93		0	
-True	255	C3	WT OD6 spoVG	11.95		0	
-True	255	C4	WT ODON spoVG	13.71		0	
-True	255	C5	WT ODON spoVG	13.87		0	
-True	255	C6	WT ODON spoVG	13.55		0	
-True	255	C7	WT OD0.5 5S	  4.91		0	
-True	255	C8	WT OD0.5 5S	  4.96		0	
-True	255	C9	WT OD0.5 5S	  5.01		0	
-True	255	C10	WT OD1.5 5S	  4.84		0	
-True	255	C11	WT OD1.5 5S	  4.78		0	
-True	255	C12	WT OD1.5 5S	  4.99		0	
-True	255	D1	WT OD3 5S	  4.93		0	
-True	255	D2	WT OD3 5S	  4.81		0	
-True	255	D3	WT OD3 5S	  4.75		0	
-True	255	D4	WT OD6 5S	  4.80		0	
-True	255	D5	WT OD6 5S	  4.67		0	
-True	255	D6	WT OD6 5S	  4.79		0	
-True	255	D7	WT ODON 5S	  4.62		0	
-True	255	D8	WT ODON 5S	  4.68		0	
-True	255	D9	WT ODON 5S	  5.00		0	
-True	65280	D10	Sample 46			0	
-True	65280	D11	Sample 47			0	
-True	65280	D12	Sample 48			0	
-True	255	E1	dAgr OD0.5 glyS	13.90		0	
-True	255	E2	dAgr OD0.5 glyS	13.94		0	
-True	255	E3	dAgr OD0.5 glyS	14.00		0	
[…]
