+++ vignettes/calibration_vignette.Rmd
@@ -123,7 +123,7 @@ display_plate(plate1plan)
 
 ```{r load_plates,dependson="label_plates",results="show"}
 plates <- read_lightcycler_1colour_cq(
-  "../inst/extdata/Edward_qPCR_Nrd1_calibration_2019-02-02_Ct.txt"
+  "../inst/extdata/Edward_qPCR_Nrd1_calibration_2019-02-02_Cq.txt.gz"
   ) %>%
   right_join(plate1plan)
 ```
@@ -237,7 +237,7 @@ ggplot(
 ```{r load_amp,dependson="label_plates",results="show"}
 
 plate1curve <- read_lightcycler_1colour_raw(
-  "../inst/extdata/Edward_qPCR_Nrd1_calibration_2019-02-02.txt"
+  "../inst/extdata/Edward_qPCR_Nrd1_calibration_2019-02-02.txt.gz"
   ) %>%
   debaseline() %>%
   left_join(plate1plan)

[…]
