### 2021-12-17 ewallace/tidyqpcr#130 @ewallace
Fix vignette titles from @wolski review
From @wolski review https://github.com/ropensci/software-review/issues/470#issuecomment-991272373

> Rmarkdown and vignette title should be the same:
> You have:
> title: "Primers and probes calibration vignette"
> %\VignetteIndexEntry{PrimerCalibration}
> but better
> title: "Primers and probes calibration vignette"
> %\VignetteIndexEntry{Primers and probes calibration vignette}
