### 2021-12-17 ewallace/tidyqpcr#131 @ewallace
Visualizing measurement results in the plate layout from @wolski review
From @wolski review https://github.com/ropensci/software-review/issues/470#issuecomment-991272373

> Would you please provide functions to visualize the results in plate layout, e.g., how those quantitation cycle, delta Cq looks on the plate? This would allow visualizing if there is a plate position effect.

I think it's useful to have a function to visualise Cq (or delta_Cq) on a plate, a bit like `display_plate`. Also easy to do.

Function should have default  argument`fill = Cq`, but allow changing that to visualise different conclusions.
