### 2021-12-17 ewallace/tidyqpcr#132 @ewallace
Documentation for plate setup avoiding edge wells
Explain how to setup plates avoiding edge wells, from @wolski review https://github.com/ropensci/software-review/issues/470#issuecomment-991272373

> If I am correct, thermocyclers are used to perform qPCR experiments. Aren't the wells in the plates' center exposed to different temperatures than on the edges? If so, how this impacts the quantification results? Does your design function takes this into account, and if not, why? Is there a theoretical option to block for the confounding of plate position?

We can do a good start on this in the documentation for `create_` functions, providing examples that avoid edge wells, e.g. `WellR = LETTERS[2:7]". That would be enough to tell users how to do it.
