### 2021-12-23 #470 @ewallace
Thank you everyone.

After a call with @jooolia and discussion with @DimmestP, we have made a plan to address these reviews. We’ve raised [a series of issue tickets on tidyqpcr repository, linked to this review ticket and also labeled rOpensci-review](https://github.com/ewallace/tidyqpcr/labels/rOpensci-review).

The plan is that in January we will start addressing those issues one by one. As we make progress there, we will request another round of feedback on this thread.
