### 2021-12-24 #470 @jooolia
Hi @wolski, thanks for your comment and for being clear that you will not continue to review here. I have removed you  as a reviewer in the issue fields and feel free to unsubscribe from the issue if you do not want to continue to receive notifications about new comments.
Thanks, Julia
