### 2021-12-24 #470 @jooolia
Thanks @ewallace the plan looks very clear to me regarding what you will tackle. Looking forward to the updates on your progress on the issues in 2022.
