### 2022-01-10 86c200e
Uses system.file to access extdata and changes example code to import compressed file
