+++ R/read_qpcr_data.R
@@ -38,7 +38,7 @@
 #' @export
 #' @seealso read_lightcycler_1colour_cq
 #'
-#' @examples read_lightcycler_1colour_raw(system.file("extdata/Edward_qPCR_Nrd1_calibration_2019-02-02.txt", 
+#' @examples read_lightcycler_1colour_raw(system.file("extdata/Edward_qPCR_Nrd1_calibration_2019-02-02.txt.gz", 
 #'                                                  package = "tidyqpcr"))
 #'
 read_lightcycler_1colour_raw <- function(
@@ -74,7 +74,7 @@ read_lightcycler_1colour_raw <- function(
 #' @export
 #' @seealso read_lightcycler_1colour_raw
 #'
-#' @examples read_lightcycler_1colour_cq(system.file("extdata/Edward_qPCR_Nrd1_calibration_2019-02-02_Ct.txt", 
+#' @examples read_lightcycler_1colour_cq(system.file("extdata/Edward_qPCR_Nrd1_calibration_2019-02-02_Ct.txt.gz", 
 #'                                                  package = "tidyqpcr"))
 #'
 read_lightcycler_1colour_cq <- function(
