+++ vignettes/multifactor_vignette.Rmd
@@ -115,15 +115,15 @@ Note that display_plate requires a column called `sample_id`, which here we had 
 # read my plates
 
 plate1 <- read_lightcycler_1colour_cq(
-  "../inst/extdata/Edward_qPCR_TxnInhibitors_HS_2018-06-15_plate1_Cq.txt.gz"
-  ) %>%
+  system.file("extdata/Edward_qPCR_TxnInhibitors_HS_2018-06-15_plate1_Cq.txt.gz",
+              package = "tidyqpcr")) %>%
   left_join(plateplan) %>%
   mutate(biol_rep = "1", plate = "1")
 
 
 plate2 <- read_lightcycler_1colour_cq(
-  "../inst/extdata/Edward_qPCR_TxnInhibitors_HS_2018-06-15_plate2_Cq.txt.gz"
-  ) %>%
+  system.file("extdata/Edward_qPCR_TxnInhibitors_HS_2018-06-15_plate2_Cq.txt.gz",
+              package = "tidyqpcr")) %>%
   left_join(plateplan) %>%
   mutate(biol_rep = "2", plate = "2")
 
@@ -272,15 +272,15 @@ ggplot(data = platesmed) +
 
 ```{r load_amp,dependson="label_plates",results="show"}
 plate1curve <- read_lightcycler_1colour_raw(
-  "../inst/extdata/Edward_qPCR_TxnInhibitors_HS_2018-06-15_plate1.txt.gz"
-  ) %>%
+  system.file("extdata/Edward_qPCR_TxnInhibitors_HS_2018-06-15_plate1.txt.gz",
+              package = "tidyqpcr")) %>%
   debaseline() %>%
   left_join(plateplan) %>%
   mutate(biol_rep = 1, plate = 1)
 
 plate2curve <- read_lightcycler_1colour_raw(
-  "../inst/extdata/Edward_qPCR_TxnInhibitors_HS_2018-06-15_plate2.txt.gz"
-  ) %>%
+  system.file("extdata/Edward_qPCR_TxnInhibitors_HS_2018-06-15_plate2.txt.gz",
+              package = "tidyqpcr")) %>%
   debaseline() %>%
   left_join(plateplan) %>%
   mutate(biol_rep = 2, plate = 2)
