+++ vignettes/deltacq_96well_vignette.Rmd
@@ -56,8 +56,8 @@ theme_set(theme_cowplot(font_size = 11) %+replace%
 wells_exclude <- c("D10","D11","D12", "H10","H11","H12") # these wells are blank
 
 plate_cq_extdata <- read_lightcycler_1colour_cq(
-  "../inst/extdata/Stuart_dAgr_glyS_spoVG_5S_individualWells_Cq.txt.gz"
-  ) %>%
+  system.file("extdata/Stuart_dAgr_glyS_spoVG_5S_individualWells_Cq.txt.gz",
+  package = "tidyqpcr")) %>%
   dplyr::filter(! well %in% wells_exclude ) %>% # exclude blank wells
   dplyr::select(well, sample_info, cq) 
 
@@ -65,8 +65,8 @@ plate_cq_extdata <- read_lightcycler_1colour_cq(
 readr::write_tsv(
   x = plate_cq_extdata %>%
     mutate(cq = format(cq,nsmall=2)),
-  path = "../inst/extdata/Stuart_dAgr_glyS_spoVG_5S_individualWells_Cq_select.txt.gz"
-)
+  path =  system.file("extdata/Stuart_dAgr_glyS_spoVG_5S_individualWells_Cq_select.txt.gz",
+                      package = "tidyqpcr")))
 
 plate_cq_extdata
 ```
