### 2022-01-11 6cd6d59
Changes Ct to Cq in read data example
