### 2022-01-11 9fe8145
Updates lightcycler Rd file
