### 2022-01-11 d2030f1
Clarifies README file
Fixes Issue #121 

- Adds explanation of MIQE guidelines in "Best-practice compliant" section of introduction
- Adds link to Rtools
- Changes links to vignettes in "Using tidyqpcr" to website rendered vignettes.
