### 2022-01-11 7f31f74
Clarifies README file (#136)
* Clarifies README file

Fixes Issue #121 

- Adds explanation of MIQE guidelines in "Best-practice compliant" section of introduction
- Adds link to Rtools
- Changes links to vignettes in "Using tidyqpcr" to website rendered vignettes.
- nudge date, flag delta Cq earlier

Co-authored-by: Edward Wallace <ewallace@users.noreply.github.com>
