### 2022-01-11 bd39b60
Compress data 128 and uses system.file 122 (#135)
* gzipped inst/extdata, read .txt.gz in vignettes, Cq replaces Ct in filenames

* Uses system.file to access extdata and changes example code to import compressed file

* Changes Ct to Cq in read data example

* Updates lightcycler Rd file

Co-authored-by: ewallace <Edward.Wallace@ed.ac.uk>
