+++ inst/extdata/Edward_qPCR_TxnInhibitors_HS_2018-06-15_plate2_Ct.txt
@@ -1,386 +1,0 @@
-Experiment: Edward qPCR Txn Inhibitors HS 2018-06-15 plate 2  Selected Filter: SYBR Green I / HRM Dye (483-533)
-Include	Color	Pos	Name	Cp	Concentration	Standard	Status
-True	255	A1	Sample 1	17.78		0	
-True	255	A2	Sample 2	16.45		0	
-True	255	A3	Sample 3	15.94		0	
-True	255	A4	Sample 4	14.95		0	
-True	255	A5	Sample 5	16.01		0	
-True	255	A6	Sample 6	16.84		0	
-True	255	A7	Sample 7	18.23		0	
-True	255	A8	Sample 8	15.66		0	
-True	255	A9	Sample 9	15.77		0	
-True	255	A10	Sample 10	14.44		0	
-True	255	A11	Sample 11	14.80		0	
-True	255	A12	Sample 12	  6.54		0	
-True	255	A13	Sample 13	16.71		0	
-True	255	A14	Sample 14	15.61		0	
-True	255	A15	Sample 15	14.89		0	
-True	255	A16	Sample 16	14.49		0	
-True	255	A17	Sample 17	14.81		0	
-True	255	A18	Sample 18	  9.80		0	
-True	255	A19	Sample 19	30.05		0	
-True	65280	A20	Sample 20			0	
-True	65280	A21	Sample 21			0	
-True	65280	A22	Sample 22			0	
-True	65280	A23	Sample 23			0	
-True	255	A24	Sample 24	30.57		0	
-True	255	B1	Sample 25	19.34		0	
-True	255	B2	Sample 26	19.10		0	
-True	255	B3	Sample 27	19.26		0	
-True	255	B4	Sample 28	15.99		0	
-True	255	B5	Sample 29	18.68		0	
-True	255	B6	Sample 30	18.51		0	
-True	255	B7	Sample 31	19.00		0	
-True	255	B8	Sample 32	22.35		0	
-True	255	B9	Sample 33	22.86		0	
-True	255	B10	Sample 34	16.54		0	
-True	255	B11	Sample 35	18.12		0	
-True	255	B12	Sample 36	17.78		0	
-True	255	B13	Sample 37	17.76		0	
-True	255	B14	Sample 38	18.70		0	
-True	255	B15	Sample 39	18.74		0	
-True	255	B16	Sample 40	15.92		0	
-True	255	B17	Sample 41	18.24		0	
-True	255	B18	Sample 42	17.95		0	
-True	65280	B19	Sample 43			0	
-True	65280	B20	Sample 44			0	
-True	65280	B21	Sample 45			0	
-True	65280	B22	Sample 46			0	
-True	65280	B23	Sample 47			0	
-True	255	B24	Sample 48	29.76		0	
-True	255	C1	Sample 49	18.94		0	
-True	255	C2	Sample 50	18.69		0	
-True	255	C3	Sample 51	18.79		0	
-True	255	C4	Sample 52	16.98		0	
-True	255	C5	Sample 53	18.22		0	
-True	255	C6	Sample 54	17.81		0	
-True	255	C7	Sample 55	19.19		0	
-True	255	C8	Sample 56	18.86		0	
-True	255	C9	Sample 57	19.07		0	
-True	255	C10	Sample 58	17.70		0	
-True	255	C11	Sample 59	17.67		0	
-True	255	C12	Sample 60	17.13		0	
-True	255	C13	Sample 61	18.96		0	
-True	255	C14	Sample 62	18.88		0	
-True	255	C15	Sample 63	18.30		0	
-True	255	C16	Sample 64	16.78		0	
-True	255	C17	Sample 65	17.80		0	
-True	255	C18	Sample 66	17.56		0	
[…]
