+++ man/read_lightcycler_1colour_cq.Rd
@@ -35,7 +35,7 @@ Lightcycler software. That software calls cq, "Cp".
 This function is a thin wrapper around readr::read_tsv.
 }
 \examples{
-read_lightcycler_1colour_cq(system.file("extdata/Edward_qPCR_Nrd1_calibration_2019-02-02_Ct.txt", 
+read_lightcycler_1colour_cq(system.file("extdata/Edward_qPCR_Nrd1_calibration_2019-02-02_Cq.txt.gz", 
                                                  package = "tidyqpcr"))
 
 }
