### 2022-01-11 64ec810
echo vignette setup fixes #137, titles fixes #130
