### 2022-01-11 dd6abd8
avoid edge wells documentation, fixes #132
