+++ vignettes/platesetup_vignette.Rmd
@@ -303,6 +303,7 @@ The functions `unite` and `separate` have visual descriptions on the [RStudio da
 
 ## Hints and Tips
 
-* In tidyqpcr we chose to give all variables snake_case names, such as `sample_id`, `prep_type`, following the tidyverse style guide.
+* In tidyqpcr we chose to give all variables snake_case names, such as `sample_id`, `prep_type`, following the [tidyverse style guide](https://style.tidyverse.org).
 * What if you are loading a 384-well plate with a fixed-spacing multichannel pipette that loads every second row? If you want to set up the plate plan to load the same sample_id/target_id in two adjacent rows, using `rep(sample_id, each = 2)` might help.
 * If your experiment spans multiple plates, it can be helpful to re-use `colkey` and `rowkey` in defining similar plate plans. You could re-use the entire plate plan if the plates are exact replicates, as long as you ensure the `sample_id` name includes replicate information and so is unique for the analysis.
+* If you want to avoid certain rows or columns, change the `well_row` or `well_col` arguments in your blank plates, row keys, and column keys. For example `create_blank_plate(well_row = LETTERS[2:7], well_col = 2:11)` creates a blank 96-well plate with outside rows and columns empty.
