### 2022-01-11 5e3550b
Create function to plot variable value across wells
