### 2022-01-12 b5150ec
Add missing Rd docs
