+++ man/make_row_names_lc1536.Rd
@@ -1,0 +1,32 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/plate_functions.R
+\name{make_row_names_lc1536}
+\alias{make_row_names_lc1536}
+\title{Generates row names for the Roche Lightcycler (tm) 1536-well plates}
+\usage{
+make_row_names_lc1536()
+}
+\value{
+Vector of row names: Aa,Ab,Ac,Ad,Ba,...,Hd.
+}
+\description{
+Creates a vector containing 36 row names according to the labeling system 
+used by the Roche Lightcycler (tm)
+}
+\examples{
+make_row_names_lc1536()
+
+}
+\seealso{
+Other plate creation functions: 
+\code{\link{create_blank_plate}()},
+\code{\link{create_colkey_4diln_2ctrl_in_24}()},
+\code{\link{create_colkey_6_in_24}()},
+\code{\link{create_colkey_6diln_2ctrl_in_24}()},
+\code{\link{create_rowkey_4_in_16}()},
+\code{\link{create_rowkey_8_in_16_plain}()},
+\code{\link{display_plate}()},
+\code{\link{label_plate_rowcol}()},
+\code{\link{make_row_names_echo1536}()}
+}
+\concept{plate creation functions}
