### 2022-01-12 ea541bf
Simplify examples and add use case in vignette
