+++ man/display_well_value.Rd
@@ -1,0 +1,51 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/plate_functions.R
+\name{display_well_value}
+\alias{display_well_value}
+\title{Display the value of each well across the plate.}
+\usage{
+display_well_value(plate, value = "cq")
+}
+\arguments{
+\item{plate}{tibble with variables well_col, well_row, and the variable to be plotted.}
+}
+\value{
+ggplot object; major output is to plot it
+}
+\description{
+Plots the plate with each well coloured by its value. Example values are Cq, Delta Cq or Delta Delta Cq.
+}
+\details{
+For a specific example see the calibration vignette:
+\code{vignette("calibration_vignette", package = "tidyqpcr")}
+}
+\examples{
+library(dplyr)
+
+# create basic 384 well plate
+basic_plate <- create_blank_plate_96well() \%>\%
+    mutate(cq = runif(96) * 10,
+    deltacq = runif(96) * 2)
+
+
+# display well Cq value across plate
+display_well_value(basic_plate)
+
+# display well Delta Cq value across plate with red colour pallette
+display_well_value(basic_plate, value = "deltacq") +   # uses ggplot syntax
+    scale_fill_gradient(high = "#FF0000") 
+          
+
+}
+\seealso{
+Other plate creation functions: 
+\code{\link{create_blank_plate}()},
+\code{\link{create_colkey_4diln_2ctrl_in_24}()},
+\code{\link{create_colkey_6_in_24}()},
+\code{\link{create_colkey_6diln_2ctrl_in_24}()},
+\code{\link{create_rowkey_4_in_16}()},
+\code{\link{create_rowkey_8_in_16_plain}()},
+\code{\link{display_plate}()},
+\code{\link{label_plate_rowcol}()}
+}
+\concept{plate creation functions}
