+++ NAMESPACE
@@ -17,6 +17,7 @@ export(create_rowkey_4_in_16)
 export(create_rowkey_8_in_16_plain)
 export(debaseline)
 export(display_plate)
+export(display_well_value)
 export(getdRdTall)
 export(label_plate_rowcol)
 export(make_row_names_echo1536)
