### 2022-01-12 e0c132d
Add ggplot2 library to example code
