+++ R/plate_functions.R
@@ -401,6 +401,7 @@ display_plate <- function(plate) {
 #'
 #' @examples 
 #' library(dplyr)
+#' library(ggplot2)
 #' 
 #' # create basic 384 well plate
 #' basic_plate <- create_blank_plate_96well() %>%
