### 2022-01-12 01002b3
Updates display_well_value Rd file
