### 2022-01-12 ewallace/tidyqpcr#126 @DimmestP
Draft response in review thread:

tidyqpcr was created to improve the quality of qPCR experiments using programming and experimental best-practices whilst being accessible to an audience that is unfamiliar with R data analysis. tidyqpcr follows the tidy data paradigm to ensure each step is human and machine readable. It also fully integrates into the wider collection of general data analysis packages provided by the tidyverse. This allows new users to get to grips with the basic pipeline quickly and expert users to develop customised analysis pipelines & visualisations.

Several other open-source libraries for qPCR analysis are available but not with the same aims as tidyqpcr. HTqPCR, ReadqPCR and NormqPCR have similar (and in some respects greater) functionality than tidyqpcr but follow an object oriented framework using S4 classes which does not align with our aims. qpcR (Spiess, 2018), which uses S3 classes not S4 as we previously stated, also uses an object oriented framework. PlateDesigner encourages better experimental design in qPCR experiments but does not support high-throughput experiments that use 384/1536 well plates. Finally, pcr (Ahmed & Kim, 2018) aligns most closely with tidyqpcr but its function inputs are not in a tidy format. Although many alternatives are available, non align with tidyqpcr’s aim to improve the quality of qPCR experiments from plate design to plotting results.
