### 2022-01-12 d048764
Make reference to values explicit
