### 2022-01-12 d26970c
Adds edge wells documentation and addresses kelshmo comments (#138)
* Adds edge wells documentation and kelshmo comments

* Add missing Rd docs
