### 2022-01-12 67cb432
Add description of value variable in docs
