+++ R/plate_functions.R
@@ -396,6 +396,8 @@ display_plate <- function(plate) {
 #' \code{vignette("calibration_vignette", package = "tidyqpcr")}
 #'
 #' @param plate tibble with variables well_col, well_row, and the variable to be plotted.
+#' 
+#' @param value character vector selecting the variable in plate to plot as the well value
 #'
 #' @return ggplot object; major output is to plot it
 #'
@@ -426,7 +428,7 @@ display_plate <- function(plate) {
 display_well_value <- function(plate, value = "cq") {
     
     # check value exists in given plate
-    assertthat::assert_that({{value}} %in% names(plate), msg = paste0({{value}}, " is not the name of a variable in the given plate"))
+    assertthat::assert_that(value %in% names(plate), msg = paste0(value, " is not the name of a variable in the given plate"))
     
     # check each well has one value only
     unique_well_value <- plate %>%
@@ -434,7 +436,7 @@ display_well_value <- function(plate, value = "cq") {
         dplyr::summarise(num_well = dplyr::n()) %>%
         dplyr::mutate(not_equal_one = .data$num_well != 1)
     
-    assertthat::assert_that(sum(unique_well_value$not_equal_one) == 0, msg = paste0("Wells do not have unique ", {{value}}, " value."))
+    assertthat::assert_that(sum(unique_well_value$not_equal_one) == 0, msg = paste0("Wells do not have unique ", value, " value."))
     
     rowlevels <- 
         dplyr::pull(plate, .data$well_row) %>%
@@ -444,7 +446,7 @@ display_well_value <- function(plate, value = "cq") {
     ggplot2::ggplot(data = plate,
                     ggplot2::aes(x = as_factor(.data$well_col),
                                  y = as_factor(.data$well_row))) +
-        ggplot2::geom_tile(ggplot2::aes(fill = .data[[{{value}}]])) +
+        ggplot2::geom_tile(ggplot2::aes(fill = .data[[value]])) +
         ggplot2::scale_x_discrete(expand = c(0, 0)) +
         ggplot2::scale_y_discrete(expand = c(0, 0),
                                   limits = rev(rowlevels)) +
