### 2022-01-18 5fe6886
document 1 plate only in calculate_drdt_plate, assert columns present, fixes #85
