### 2022-01-18 ccf150c
remove getdRdTall, fixes #103
