### 2022-01-18 ewallace/tidyqpcr#126 @ewallace
My edits included below:

tidyqpcr was created to improve the quality of qPCR experimental design and analysis whilst being accessible to an audience that is new to R data analysis. Every data object in tidyqpcr is a tidy data frame, which makes the data both human and machine readable, enabling new users to get to grips with tidyqpcr features quickly. Following the tidy data paradigm integrates tidyqpcr into the wider collection of data analysis packages provided by the tidyverse. Tidyverse integration enables users to develop customised analysis pipelines and visualisations including integrating with other tidy data.

Several other open-source libraries for qPCR analysis are available with distinct aims. HTqPCR (Dvinge 2009), ReadqPCR/NormqPCR (Perkins et al., 2012), and qpcR (Spiess, 2018) have similar and in some respects greater functionality than tidyqpcr, but follow object oriented approaches which do not align with our aims. PlateDesigner encourages better experimental design in qPCR experiments but does not support 384- or 1536-well plates. Finally, pcr (Ahmed & Kim, 2018) aligns most closely with tidyqpcr but its function inputs are not tidy data frames. Although many alternatives are available, tidyqpcr's aims and approach are distinct: to improve the quality of qPCR experiments from plate design to analysis, by exposing all data in a consistent tidy format that integrates with the tidyverse.
