### 2022-01-18 ewallace/tidyqpcr#126 @ewallace
@DimmestP let's discuss the paper.md file too.
