### 2022-01-24 16630d2
Rewrite description of other packages, fixes #126
