### 2022-01-27 4d9a51e
Added MIQE guideline comparison table
Adds draft table in README below software features to compare tidyqpcr to other current software according to compliance with MIQE guidelines. 

Current tidyqcr is compared to pcr, HTqPCR and NormqPCR. I have not yet finished comparison with qpcR as the documentation is a little convoluted and I need more time to find out exactly what is being done.

There remains a few question marks in the tidyqpcr column as I am not entirely sure if we fully complete the requirements by MIQE on those topics.
