+++ vignettes/calibration_vignette.Rmd
@@ -262,7 +262,7 @@ ggplot(
 file_path_raw <- system.file("extdata/Edward_qPCR_Nrd1_calibration_2019-02-02.txt.gz",
                              package = "tidyqpcr")
 
-plate1curve <- file_path_amp %>%
+plate1curve <- file_path_raw %>%
     read_lightcycler_1colour_raw() %>%
     debaseline() %>%
     left_join(plate1plan)
