### 2022-02-03 de13785
Adds missing references
