+++ paper.bib
@@ -51,6 +51,40 @@
     note = {R package version 1.4-1},
     url = {https://CRAN.R-project.org/package=qpcR},
   }
+  
+@article{Perkins:2012,
+   abstract = {Background: Measuring gene transcription using real-time reverse transcription polymerase chain reaction (RT-qPCR) technology is a mainstay of molecular biology. Technologies now exist to measure the abundance of many transcripts in parallel. The selection of the optimal reference gene for the normalisation of this data is a recurring problem, and several algorithms have been developed in order to solve it. So far nothing in R exists to unite these methods, together with other functions to read in and normalise the data using the chosen reference gene(s).Results: We have developed two R/Bioconductor packages, ReadqPCR and NormqPCR, intended for a user with some experience with high-throughput data analysis using R, who wishes to use R to analyse RT-qPCR data. We illustrate their potential use in a workflow analysing a generic RT-qPCR experiment, and apply this to a real dataset. Packages are available from http://www.bioconductor.org/packages/release/bioc/html/ReadqPCR.htmland http://www.bioconductor.org/packages/release/bioc/html/NormqPCR.html. Conclusions: These packages increase the repetoire of RT-qPCR analysis tools available to the R user and allow them to (amongst other things) read their data into R, hold it in an ExpressionSet compatible R object, choose appropriate reference genes, normalise the data and look for differential expression between samples. © 2012 Perkins et al.; licensee BioMed Central Ltd.},
+   author = {James R. Perkins and John M. Dawes and Steve B. McMahon and David L.H. Bennett and Christine Orengo and Matthias Kohl},
+   doi = {10.1186/1471-2164-13-296/FIGURES/4},
+   issn = {14712164},
+   issue = {1},
+   journal = {BMC Genomics},
+   keywords = {Animal Genetics and Genomics,Life Sciences,Microarrays,Microbial Genetics and Genomics,Plant Genetics and Genomics,Proteomics,general},
+   month = {7},
+   pages = {1-8},
+   pmid = {22748112},
+   publisher = {BioMed Central},
+   title = {ReadqPCR and NormqPCR: R packages for the reading, quality checking and normalisation of RT-qPCR quantification cycle (Cq) data},
+   volume = {13},
+   url = {https://bmcgenomics.biomedcentral.com/articles/10.1186/1471-2164-13-296},
+   year = {2012},
+}
+  
+@article{Dvinge:2009,
+    author = {Dvinge, Heidi and Bertone, Paul},
+    title = "{HTqPCR: high-throughput analysis and visualization of quantitative real-time PCR data in R}",
+    journal = {Bioinformatics},
+    volume = {25},
+    number = {24},
+    pages = {3325-3326},
+    year = {2009},
+    month = {10},
+    abstract = "{Motivation: Quantitative real-time polymerase chain reaction (qPCR) is routinely used for RNA expression profiling, validation of microarray hybridization data and clinical diagnostic assays. Although numerous statistical tools are available in the public domain for the analysis of microarray experiments, this is not the case for qPCR. Proprietary software is typically provided by instrument manufacturers, but these solutions are not amenable to the tandem analysis of multiple assays. This is problematic when an experiment involves more than a simple comparison between a control and treatment sample, or when many qPCR datasets are to be analyzed in a high-throughput facility.Results: We have developed HTqPCR, a package for the R statistical computing environment, to enable the processing and analysis of qPCR data across multiple conditions and replicates.Availability:HTqPCR and user documentation can be obtained through Bioconductor or at http://www.ebi.ac.uk/bertone/software.Contact:bertone@ebi.ac.uk}",
[…]
