### 2022-02-03 577eb2f
Clarify MIQE table with LOD etc.
