### 2022-02-04 #470 @ewallace
Hello @jooolia and @kelshmo!

After a lot of work, @DimmestP and I are resubmitting our improved tidyqpcr package for review. We created a new tag [tidyqpcr v0.4](https://github.com/ewallace/tidyqpcr/releases/tag/v0.4), for the latest version.

The issues we addressed are listed in the tidyqpcr repo under the [rOpensci-review label](https://github.com/ewallace/tidyqpcr/issues?q=label%3ArOpensci-review+).

We did not address [new vignette on normalisation functions tidyqpcr#127](https://github.com/ewallace/tidyqpcr/issues/127), because we decided it was not essential and so prioritised other things.

We look forward to your feedback and review.
