### 2022-02-19 a9d60c8
Clarify 96-well vignette, sample_info, target_id
- don't group by sample_info
- include target_id in workflow summary
- clarify text at other points
