+++ vignettes/deltacq_96well_vignette.Rmd
@@ -110,7 +110,7 @@ The `sample_info` column contains values like `WT OD0.5 glyS`, where `WT` descri
 
 However, tidyqpcr expects a column called `sample_id` that identifies the unique sample - here, that would be identified by a strain (e.g. `WT`) and growth stage (e.g. `OD0.5`). So we use the `unite` function, from the `tidyr` package, to create a new column `sample_id` that unites strain and OD.
 
-Before these, we have to do a little organization. Wee add row and column labels for each well, so that we can later draw the plate plan and check the description. We use the tidyqpcr function `create_blank_plate_96well` to make a "blank" plate plan, then join that information to the data. The join function `inner_join` is in the `dplyr` package.
+Before these, we have to do a little organization. We add row and column labels for each well, so that we can later draw the plate plan and check the description. We use the tidyqpcr function `create_blank_plate_96well` to make a "blank" plate plan, then join that information to the data. The join function `inner_join` is in the `dplyr` package.
 
 We use the tidyverse pipe operator `%>%` to do all these transformations in a row, with one on each line.
 
@@ -208,7 +208,7 @@ Having made normalized values, we use `group_by` and `summarize` from the `dplyr
 
 ```{r median_counts,dependson="load_plates"}
 plate_med <- plate_norm %>%
-  group_by(sample_info, sample_id, strain, OD, target_id) %>%
+  group_by(sample_id, strain, OD, target_id) %>%
   summarize(
     delta_cq  = median(delta_cq, na.rm = TRUE),
     rel_abund = median(rel_abund, na.rm = TRUE)
@@ -290,7 +290,7 @@ plate_deltanorm <- plate_norm %>%
   calculate_deltadeltacq_bytargetid(ref_sample_ids = "WT_OD0.5")
 
 plate_deltanorm_med <- plate_deltanorm %>%
-    group_by(sample_info, sample_id, strain, OD, target_id) %>%
+    group_by(sample_id, strain, OD, target_id) %>%
   summarize(
     deltadelta_cq  = median(deltadelta_cq, na.rm = TRUE),
     fold_change    = median(fold_change,   na.rm = TRUE)
@@ -338,8 +338,9 @@ ggplot(data = filter(plate_deltanorm_med, target_id != "5S") ) +
 In this vignette, you have learned the basic workflow for analysing Cq values from qPCR data: 
 
 * Create the `sample_id` column to uniquely identify each biological sample.
-* Calculate delta Cq and relative abundance using `calculate_deltacq_bysampleid`. This process normalizes target gene values to the values of one or more reference targets within each sample.
-* Calculate delta delta Cq and fold change using `calculate_deltadeltacq_bytargetid`. This process normalizes each target gene value for each sample to values of the same target in one or more reference samples. 
+* Create the `target_id` column to uniquely identify each target amplicon, e.g. detecting one gene.
+* Calculate delta Cq and relative abundance using `calculate_deltacq_bysampleid`. This process normalizes values for each target to the values of one or more reference targets within each sample.
+* Calculate delta delta Cq and fold change using `calculate_deltadeltacq_bytargetid`. This process normalizes the value for each target in each sample to values of the same target in one or more reference samples. 
 
 # Notes on normalization, reference targets, and reference samples
 
@@ -347,10 +348,23 @@ In this section, we discuss the process of normalization and the choices of refe
 
 Normalization is a tool for making specific comparisons in specific experimental designs.
 
[…]
