### 2022-02-19 e80b554
Explain factors in vignettes, improve camibration vis
