+++ vignettes/platesetup_vignette.Rmd
@@ -26,7 +26,9 @@ Setting up plates is partly a technical question of how to use functions in tidy
 
 We suggest thinking through the whole experiment first, including what you will measure, how many replicates, and what figures you will want to make. If you plan all the analysis before even starting to grow your biological samples and extract RNA/DNA, then it is easier to avoid mistakes. Also, the steps from cell growth, through nucleic acid extraction and measurement, to finished figures, go much quicker.
 
-This vignette builds from a 12-well "practice plate" up to a 96 well plate for a plausible small RT-qPCR experiment. The goal is that after working through this vignette, the plate setup in the Multifactorial vignette will be easier to follow.
+This vignette builds from a 12-well "practice plate" up to a 96 well plate for a plausible small RT-qPCR experiment.
+The key idea is to design one replicate of your experiment in a small rectangle on the plate, then make copies of this small rectangle across the plate for more replicates or more complicated designs.
+The goal is that after working through this vignette, the plate setup in the Multifactorial vignette will be easier to follow.
 
 This vignette focuses on one primer set per well (for SYBR data), and doesn't discuss more than one probe per well (for TaqMan data). Please create [an issue on the github](https://github.com/ewallace/tidyqpcr/issue) if your data needs aren't covered here, and we will try to respond to it.
 
@@ -150,26 +152,29 @@ It also means that technical replicates of the same sample are not in adjacent w
 However, if there are row- or column-specific artefacts, this approach does not allow you detect them separately.
 
 
-## Repeat columns for the sample_ids and prep_types
+## Replicate columns for the sample_ids and prep_types
 
-Here we are putting three replicates of +RT and one of -RT for each sample. This approach is reliable if DNA contamination from -RT samples would show up in multiple sample/target combinations. 
+Here we are putting three technical replicates of +RT and one of -RT for each sample.
+This approach is reliable if DNA contamination from -RT samples would show up in multiple sample/target combinations. 
 
-We could achieve these replicates in the plate plan by explicitly writing out every time as in `c("+RT", "+RT", "+RT", "-RT")`, or we can again use the `rep` function. Below, we use `rep("+RT", times = 9)` to make 9 repeats, meaning that the 3 tech reps of +RT samples are next to each other. We use the concatenate function `c`, to arrange that next to the single replicates of the -RT samples. 
+We could achieve these replicates in the plate plan by explicitly writing out every time as in `c("+RT", "+RT", "+RT", "-RT")`, or we can again use the `rep` function.
+Below, we use `rep("+RT", times = 9)` to make 9 repeats, meaning that the 3 technical replicates of each of 3 +RT samples are next to each other. 
+We use the concatenate function `c`, to arrange that next to the single replicates of the 3 -RT samples. 
 
 ```{r minimal_48_cols,results="show",echo=TRUE}
 sample_id_levels <- c("rep1", "rep2", "rep3")
-prep_type_vector <- c(rep("+RT", times = 9), rep("-RT", times = 3))
-print(prep_type_vector)
+prep_type_values <- c(rep("+RT", times = 9), rep("-RT", times = 3))
+print(prep_type_values)
 
 colkey12 <- tibble(
   well_col = 1:12,
   sample_id = rep(sample_id_levels, times = 4),
-  prep_type = prep_type_vector
+  prep_type = prep_type_values
 )
 print(colkey12)
 ```
 
-## Putting the 48-rep sample together
+## Putting the 48-well sample together
 
 ```{r minimal_48_combine,results="show",echo=TRUE}
 
@@ -191,7 +196,8 @@ display_plate(plate_plan48)
 
 # Adding more samples and repeating targets
 
[…]
