### 2022-03-18 ewallace/tidyqpcr#125 @jooolia
Hello, I am looking through the changes and wanted to note that this table is very useful and clarifies the functionality of tidyqpcr compared to other packages.
