### 2022-03-18 #470 @jooolia
Hello @ewallace , I have been in contact with Kelsey who will have a look soon. I am having a look through the responses to the reviews and then will have a through the overall package. Apologies this has been going a bit more slowly than I expected (we rotate EiC duties and I am currently editor in chief at rOpenSci).
