### 2022-04-01 #470 @kelshmo
@ewallace Great work, thanks for your patience in working with me. Some communication early on fell through the cracks when I switched job roles!

Your addition of documentation and revised function names greatly helps with package clarity. I confirmed your use of file paths relative to the user function well. Your README and github pages look great. approved! Anything else you need from me?
