### 2022-04-15 #470 @jooolia
Dear @ewallace,
Thank you for your patience with my review.

This is part 1/2 of my review and will concern mostly aspects related to packaging. Part 2/2 will look more into the functions and the code (and more at the tests).

## Check package integrity

### run checks on `tidyqpcr` source:

```
devtools::check(pkg_dir)
```

This is mostly ok, but there are several files that should be added to your .Rbuildignore. ( ^codemeta\.json$, ^CITATION\.cff$, ^bibtex\.bib$, ^paper\.md$) so that you do not get notes when running the checks.

In your DESCRIPTION, the Apache license is one accepted by CRAN and rOpenSci, but  it should written "Apache License (>= 2)" instead, you can use usethis::use_apache_license() to add the appropriate text to your DESCRIPTION and the file to your directory (and this will add LICENSE.md to .Rbuildignore). (Funnily enough it seems that Apache License (>=2) + file LICENSE is not ok, but MIT + file LICENSE doesn't raise any issues with check()).

### run tests on `tidyqpcr` source:

```{r check-tests}
devtools::test(pkg_dir)
```

Running tests using devtools::test() several tests fail (test-amp_melt_curve_functions.R:5:1,test-calculate_deltacq.R:5:1, test-plate_functions.R:6:1) because the needed packages are not loaded and not specified using ::, for example mutate and pull do not have references to dplyr.

## Check package metadata files

### Readme

The Readme is very comprehensive and provides a lot of information. For impatient people like me would you consider moving the Getting Started section closer to the top?

There is no code in the Readme so I think it is ok to have it as a .md.

### Description

I see that on Github the version is 0.4 but in the description it is 0.3. Should this be updated? If you had different versions I would highly  recommend a NEWS.md file to briefly describe the updates in the version (not requesting that you do this retrospectively, but would help with these changes).

Would consider moving tibble and tidyr to "imports" instead of "depends"? Or is there a specific reason for this?

## Check documentation

### test `tidyqpcr` function help files:

No top-level help-file exists. Please add one, more info here: https://devguide.ropensci.org/building.html#documentation

## Review test suite:

### test coverage

Currently no tests for two functions in read_qpcr_data.R and I think these are  important functions for the package that need to be tested.

As I mentioned above this is a first part of the review. I will be away for Easter break, but will start working on part 2 upon my return.

Thanks, Julia
