### 2022-04-15 ewallace/tidyqpcr#152 @ewallace
Integrity and metadata @jooolia review part 1
From https://github.com/ropensci/software-review/issues/470#issuecomment-1100386014

- [x] build .Rbuildignore
- [x] fix LICENSE.md
- [x] load packages in tests
- [x] move "getting started" further up in README.md
- [x] DESCRIPTION, import tidyr and fix version 0.4. (kept Depends on tibble because it makes the examples nicer)
- [x] top-level help file
