### 2022-04-15 ewallace/tidyqpcr#153 @ewallace
Tests for read data functions
From @jooolia review part 1 https://github.com/ropensci/software-review/issues/470#issuecomment-1100386014

Add tests for functions in `read_qpcr_data.R`.
