### 2022-04-15 fa29569
update DESCRIPTION, format license, ignore more stuff
