### 2022-04-16 f5816f1
short tests for read functions
