### 2022-04-16 ewallace/tidyqpcr#157 @ewallace
Add devtools::test() to continuous integration
From https://github.com/ropensci/software-review/issues/470#issuecomment-1100386014

@jooolia pointed out that `devtools::test()` was failing. I had not until that point realised that `devtools::check()` did not run the tests.

Could we add `devtools::test()` to the continuous integration on github, to make sure that doesn't happen again?
