+++ vignettes/calibration_vignette.Rmd
@@ -46,17 +46,13 @@ knitr::opts_chunk$set(
 
 # Load packages
 library(tidyverse)
-library(cowplot)
 library(tidyqpcr)
 
 # set default theme for graphics
-theme_set(theme_cowplot(font_size = 11) %+replace%
+theme_set(theme_bw(base_size = 11) %+replace%
   theme(
-    panel.border = element_rect(
-      colour = "grey50",
-      linetype = "solid", size = 0.5
-    ),
-    strip.background = element_blank()
+    strip.background = element_blank(),
+    panel.grid = element_blank()
   ))
 ```
 
@@ -193,7 +189,6 @@ ggplot(data = plates) +
     title = "All reps, unnormalized"
   ) +
   facet_wrap(~biol_rep) +
-  panel_border() +
   scale_y_continuous(breaks = seq(from = 15, to = 35, by = 5)) + 
   theme(axis.text.x = element_text(angle = 90, vjust = 0.5),
         panel.grid.major.y = element_line(colour="grey80", size = 0.2))
