+++ DESCRIPTION
@@ -21,7 +21,6 @@ Suggests:
     knitr,
     rmarkdown,
     tidyverse,
-    cowplot,
     testthat,
     covr
 VignetteBuilder: knitr
