+++ vignettes/calibration_vignette.Rmd
@@ -1,7 +1,7 @@
 ---
 title: "Primers and probes calibration vignette"
 author: "Edward Wallace"
-date: "Feb 2019"
+date: "April 2022"
 output: 
   rmarkdown::html_vignette:
     toc: true
@@ -45,7 +45,9 @@ knitr::opts_chunk$set(
 )
 
 # Load packages
-library(tidyverse)
+library(tidyr)
+library(ggplot2)
+library(dplyr)
 library(tidyqpcr)
 
 # set default theme for graphics
