+++ DESCRIPTION
@@ -20,7 +20,6 @@ Imports:
 Suggests: 
     knitr,
     rmarkdown,
-    tidyverse,
     testthat,
     covr
 VignetteBuilder: knitr
