### 2022-05-01 #470 @jooolia
-  Very patient and  very open to useful suggestions from reviewers.

- [X]  As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work (if you are unsure whether you are in conflict, please speak to your editor *before* starting your review).
    -  A note:I am the editor for this package, but have also provided a review as one of the reviewers declined to continue with the process.

    - Tests are present, but are failing so it is currently difficult to assess the full extent of how well the tests are covering the package.

-   [X] Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer (“rev” role) in the package DESCRIPTION file.

Description :
* does {lifecycle} need to be in imports? It comes up as an R CMD check NOTE that it is not used.
* Do tidyr and tiblble need to be in “Depends”, could they instead be in the “Imports”?

#### Spell check

Selected results found using `devtools::spell_check()`
* designe README.md:139 - designed
* behavior calibration_vignette.Rmd:171 - elsewhere as UK spelling
* paramters platesetup_vignette.Rmd:306 - parameters
* pcr README.md:115 - PCR could be consistently capitalised here.
* targetss calculate_deltacq_bysampleid.Rd:17 - targets

#### Comments on specific functions

-   `calculate_dydx_1()` - why _1 ?

-   `label_plate_rowcol()` - I can’t find an example that works without coercefactors=TRUE, so is it necessary as an argument? I found it a bit confusing since in the plate creation you are making the plate row and columns as factors. I also found the warning about the transformation to factors confusing as I thought that I had done something wrong when constructing the plate.

-   `display_plate(create_blank_plate_1536well())` does not work which I can understand. Is there anything planned or could it just be done by colour after a certain size to show a bit the planning of a plate? Not sure this is useful.

-   In the examples in display_plate and display_well_value there are calls to “library(dplyr)”, I think this would be better as dplyr::.

-   What about the functions in plot_helpers? Where should they be used? It is not clear to me how they integrate in the package even though I could imagine they would be useful for some of the plots one may want to produce with these data. The code formatting and styling could be improved in this file (plot_helpers.R).

-   Overall the naming of functions was very consistent and made it easy to understand the functions.

-   Would it be possible to add very brief and basic usage to the Readme? I think that the vignette "Delta Cq 96-well plate qPCR analysis example" is a nice example of a workflow that anyone with qPCR data could use so maybe something basic could be adapted from this vignette?

-   The Pkgdown website is not up to date, but that should be taken care of once the package is accepted by rOpenSci.

#### Overall comments

- Nice use of @family for the plate functions.

- Is there a reason that the data sets have been kept in inst/extdata and not as datasets in data/ ? Just curious. One reason I could see is that you want to keep the original file formats as examples for the users. However the disadvantage is that the loading is slightly less easy for users (but perhaps this is negligible) and the datasets are not documented in an R/data.R file. This is not a request, but just a question for you. :)
- I also found the length of the functions to be very manageable and I think this sets the package up nicely for new additions and any desired improvements in the future.

Thanks for your patience with this process. I think the easiest is that once the addressable comments are completed that you could add a comment to this issue with the different commits used to reply to the different issues (for example something like[ this](https://github.com/ropensci/software-review/issues/498#issuecomment-1062707351 )).

Thanks, Julia
