### 2022-05-01 da45200
Remove lifecycle dependency as not used
