+++ DESCRIPTION
@@ -15,7 +15,6 @@ Imports:
     readr,
     forcats,
     assertthat,
-    lifecycle,
     tidyr
 Suggests: 
     knitr,
