### 2022-05-02 33368a4
Updates all actions to v2
setup-r-dependencies@v1 -> setup-r-dependencies@v2
