### 2022-05-02 70678eb
Updates to latest howfairis-github-action version (#160)
* Updates to latest howfairis-github-action version

Changes version of 0.2.0 to 0.2.1

* Remove ubuntu-latest devel test

Co-authored-by: Edward Wallace <ewallace@users.noreply.github.com>
