+++ .github/workflows/fair-software.yml
@@ -11,7 +11,7 @@ jobs:
     name: "fair-software"
     runs-on: ubuntu-latest
     steps:
-      - uses: fair-software/howfairis-github-action@0.2.0
+      - uses: fair-software/howfairis-github-action@0.2.1
         name: Measure compliance with fair-software.eu recommendations
         env:
           PYCHARM_HOSTED: "Trick colorama into displaying colored output" 
