+++ .github/workflows/R-CMD-check.yaml
@@ -22,7 +22,6 @@ jobs:
           - {os: windows-latest, r: 'release'}
           - {os: ubuntu-latest,   r: 'release'}
           - {os: ubuntu-latest,   r: 'oldrel-1'}
-          - {os: ubuntu-latest,   r: 'devel', http-user-agent: 'release'}
           
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
