### 2022-05-02 #470 @ewallace
Thank you @jooolia! We will work on these in some issue tickets and get back to you as soon as we can.

@kelshmo, may we also acknowledge you as a package reviewer in the DESCRIPTION file?
