### 2022-05-03 ewallace/tidyqpcr#157 @DimmestP
Yes, devtools::check() does not run the tests. However, the continuous integration on github does already successfully run the tests. The difference is that I use testthat::test_package() not devtools::test() which run slightly differently. testthat::test_package() first finds the testthat.R file where all the libraries are loaded before running the tests. devtools::test() skips this step on just runs the tests without loading the libraries. It's a relatively minor point but being explicit in the tests themselves solves the problem.
