### 2022-05-03 ewallace/tidyqpcr#162 @ewallace
Rename calculate_dydx
Rename `calculate_dydx_1` to `calculate_dydx`. It works on vectors, had the suffix `_1` to be on only 1 well of a plate, but we don't need that because `calculate_drdt_plate` already makes clear that it applies to a plate.

Per @jooolia review: https://github.com/ropensci/software-review/issues/470#issuecomment-1114160848
