### 2022-05-03 ewallace/tidyqpcr#163 @ewallace
Rethink coercing factors in label_plate_rowcol
As @jooolia said in review https://github.com/ropensci/software-review/issues/470#issuecomment-1114160848

> label_plate_rowcol() - I can’t find an example that works without coercefactors=TRUE, so is it necessary as an argument? I found it a bit confusing since in the plate creation you are making the plate row and columns as factors. I also found the warning about the transformation to factors confusing as I thought that I had done something wrong when constructing the plate.

@DimmestP reports that users have also seen the warning and worried that they were doing something wrong. We need to rethink and change this.

Ideas:
- stop the warning altogether
- give a note that isn't a warning
- go through examples so that none of them produce warnings
