### 2022-05-03 ewallace/tidyqpcr#164 @ewallace
Split display_plate into 2 functions
Responding to @jooolia's review https://github.com/ropensci/software-review/issues/470#issuecomment-1114160848

> `display_plate(create_blank_plate_1536well())` does not work which I can understand. Is there anything planned or could it just be done by colour after a certain size to show a bit the planning of a plate? Not sure this is useful.

@DimmestP and I discussed. Suggestion is to split current `display_plate` into 2 functions that have a single help page
- `display_plate` will just do the blank plate with formatting, and have at least one example with flexible colour (e.g. `+ geom_tile(aes(fill = well_row, alpha = well_col))`.
- `display_plate_qpcr` will call `display_plate` and add the specific formatting for `geom_tile` and `geom_text` involving `target_id`, `sample_id`, `prep_type`.
- remember to update vignettes with `display_plate_qpcr`.
