### 2022-05-04 45649a1
rename calculate_dydx, fixes #162
