+++ man/calculate_dydx.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/amp_melt_curve_functions.R
-\name{calculate_dydx_1}
-\alias{calculate_dydx_1}
+\name{calculate_dydx}
+\alias{calculate_dydx}
 \title{Calculate dy/dx vector from vectors y and x}
 \usage{
-calculate_dydx_1(x, y, method = "spline", ...)
+calculate_dydx(x, y, method = "spline", ...)
 }
 \arguments{
 \item{x}{input variable, numeric vector, assumed to be temperature}
@@ -35,13 +35,13 @@ y = x^2
 
 # calculate gradient of curve
 #----- use case 1 : using splines
-calculate_dydx_1(x, y)
+calculate_dydx(x, y)
 
 # optional arguments are passed to smooth.splines function
-calculate_dydx_1(x, y, spar = 0.5)
+calculate_dydx(x, y, spar = 0.5)
 
 #----- use case 2 : using difference between adjacent points
-calculate_dydx_1(x, y, method = "diff")
+calculate_dydx(x, y, method = "diff")
 
 }
 \seealso{
