### 2022-05-04 ewallace/tidyqpcr#165 @DimmestP
Improve plot_helpers.R styling
Issue highlighted by Julia in rOpenSci review https://github.com/ropensci/software-review/issues/470
