### 2022-05-04 55f4ca5
Completes styling edits according to lintr
