### 2022-05-04 bd482f2
Creates display_plate_qpcr function. Update vignettes, docs and tests
