+++ man/display_plate_qpcr.Rd
@@ -1,0 +1,49 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/plate_functions.R
+\name{display_plate_qpcr}
+\alias{display_plate_qpcr}
+\title{Display qPCR plate plan with sample_id, target_id, prep_type per well}
+\usage{
+display_plate_qpcr(plate)
+}
+\arguments{
+\item{plate}{tibble with variables well_col, well_row, sample_id, target_id,
+prep_type. Output from label_plate_rowcol.}
+}
+\value{
+ggplot object; major output is to plot it
+}
+\description{
+Display qPCR plate plan with sample_id, target_id, prep_type per well
+}
+\examples{
+library(dplyr)
+
+# create basic 384 well plate
+basic_plate <- label_plate_rowcol(create_blank_plate(), 
+                                  create_rowkey_8_in_16_plain(target_id = c("T_1", "T_2",
+                                                                            "T_3", "T_4", 
+                                                                            "T_5", "T_6",
+                                                                            "T_7", "T_8")), 
+                                  create_colkey_6diln_2ctrl_in_24() \%>\% 
+                                      mutate(sample_id = paste0(dilution_nice, "_", tech_rep)))
+
+# display full plate
+display_plate_qpcr(basic_plate)
+
+}
+\seealso{
+Other plate creation functions: 
+\code{\link{create_blank_plate}()},
+\code{\link{create_colkey_4diln_2ctrl_in_24}()},
+\code{\link{create_colkey_6_in_24}()},
+\code{\link{create_colkey_6diln_2ctrl_in_24}()},
+\code{\link{create_rowkey_4_in_16}()},
+\code{\link{create_rowkey_8_in_16_plain}()},
+\code{\link{display_plate}()},
+\code{\link{display_well_value}()},
+\code{\link{label_plate_rowcol}()},
+\code{\link{make_row_names_echo1536}()},
+\code{\link{make_row_names_lc1536}()}
+}
+\concept{plate creation functions}
