### 2022-05-04 25bed2d
Adds example code to Using tidyqpcr section
