### 2022-05-04 c07b43c
display_plate used in display_plate_value
- display_plate has a more basic basic_plate example
- display_well_value renamed to diaplay_plate_value
- display_plate now used in display_plate_value
- code styling in display_plate_...
