+++ R/plate_functions.R
@@ -385,7 +385,14 @@ label_plate_rowcol <- function(plate,
 #' @examples
 #' library(ggplot2)
 #' 
-#' # display wells of empty plate
+#' # display empty plot of empty plate
+#' display_plate(create_blank_plate_96well())
+#' 
+#' # display wells of empty plate filled by column
+#' display_plate(create_blank_plate_96well()) + 
+#'   geom_tile(aes(fill = well_col), colour = "black")
+#' 
+#' # display wells of empty 1536-well plate filled by row
 #' display_plate(create_blank_plate_1536well()) + 
 #'   geom_tile(aes(fill = well_row), colour = "black")
 #' 
@@ -424,19 +431,32 @@ display_plate <- function(plate) {
 #' @return ggplot object; major output is to plot it
 #'
 #' @examples 
-#' library(dplyr)
 #' 
-#' # create basic 384 well plate
-#' basic_plate <- label_plate_rowcol(create_blank_plate(), 
+#' 
+#' # create basic 6-well plate
+#' basic_plate <- 
+#'     label_plate_rowcol(plate = create_blank_plate(well_row = LETTERS[1:2],
+#'                                                   well_col = 1:3),
+#'                        rowkey = tibble(well_row = factor(LETTERS[1:2]),
+#'                                        target_id = c("T_A","T_B")),
+#'                        colkey = tibble(well_col = factor(1:3),
+#'                                        sample_id = c("S_1","S_2", "S_3"),
+#'                                        prep_type = "+RT"))
+#' 
+#' # display basic plate
+#' display_plate_qpcr(basic_plate)
+#' 
+#' # create full 384 well plate
+#' full_plate <- label_plate_rowcol(create_blank_plate(), 
 #'                                   create_rowkey_8_in_16_plain(target_id = c("T_1", "T_2",
 #'                                                                             "T_3", "T_4", 
 #'                                                                             "T_5", "T_6",
 #'                                                                             "T_7", "T_8")), 
 #'                                   create_colkey_6diln_2ctrl_in_24() %>% 
-#'                                       mutate(sample_id = paste0(dilution_nice, "_", tech_rep)))
+#'                                       dplyr::mutate(sample_id = paste0(dilution_nice, "_", tech_rep)))
 #' 
 #' # display full plate
-#' display_plate_qpcr(basic_plate)
+#' display_plate_qpcr(full_plate)
 #' 
 #' @family plate creation functions
 #'
@@ -474,17 +494,17 @@ display_plate_qpcr <- function(plate) {
 #' library(dplyr)
 #' library(ggplot2)
 #' 
-#' # create basic 384 well plate
-#' basic_plate <- create_blank_plate_96well() %>%
+#' # create 96 well plate with random values
+#' plate_randomcq <- create_blank_plate_96well() %>%
 #'     mutate(cq = runif(96) * 10,
-#'     deltacq = runif(96) * 2)
+#'            deltacq = runif(96) * 2)
 #' 
 #' 
 #' # display well Cq value across plate
-#' display_well_value(basic_plate)
+#' display_plate_value(plate_randomcq)
 #' 
 #' # display well Delta Cq value across plate with red colour pallette
-#' display_well_value(basic_plate, value = "deltacq") +   # uses ggplot syntax
+#' display_plate_value(plate_randomcq, value = "deltacq") +   # uses ggplot syntax
 #'     scale_fill_gradient(high = "#FF0000") 
 #'           
 #' 
@@ -494,10 +514,10 @@ display_plate_qpcr <- function(plate) {
 #' @importFrom forcats as_factor
 #' @importFrom rlang .data
 #'
-display_well_value <- function(plate, value = "cq") {
-    
+display_plate_value <- function(plate, value = "cq") {
[…]
