+++ man/create_colkey_4diln_2ctrl_in_24.Rd
@@ -46,8 +46,8 @@ Other plate creation functions:
 \code{\link{create_rowkey_4_in_16}()},
 \code{\link{create_rowkey_8_in_16_plain}()},
 \code{\link{display_plate_qpcr}()},
+\code{\link{display_plate_value}()},
 \code{\link{display_plate}()},
-\code{\link{display_well_value}()},
 \code{\link{label_plate_rowcol}()},
 \code{\link{make_row_names_echo1536}()},
 \code{\link{make_row_names_lc1536}()}
