+++ NAMESPACE
@@ -18,7 +18,7 @@ export(create_rowkey_8_in_16_plain)
 export(debaseline)
 export(display_plate)
 export(display_plate_qpcr)
-export(display_well_value)
+export(display_plate_value)
 export(label_plate_rowcol)
 export(make_row_names_echo1536)
 export(make_row_names_lc1536)
