### 2022-05-05 8bafc89
Adds missing function documentation
