+++ R/plate_functions.R
@@ -453,7 +453,9 @@ display_plate <- function(plate) {
 #'                                                                             "T_5", "T_6",
 #'                                                                             "T_7", "T_8")), 
 #'                                   create_colkey_6diln_2ctrl_in_24() %>% 
-#'                                       dplyr::mutate(sample_id = paste0(dilution_nice, "_", tech_rep)))
+#'                                       dplyr::mutate(sample_id = paste0(dilution_nice,
+#'                                                                        "_",
+#'                                                                        tech_rep)))
 #' 
 #' # display full plate
 #' display_plate_qpcr(full_plate)
