### 2022-05-05 4ac3ee1
Adds display_plate_value.Rd
