### 2022-05-06 111bcec
message for names in label_plate_rowcol, assert in display_plate_*
