### 2022-05-06 ce8f27b
assert necessary variable names in dataframes
