### 2022-05-07 b8d595a
clean up assert statements
