### 2022-05-17 ewallace/tidyqpcr#157 @ewallace
I am happy that this is solved by being explicit in the tests so that we do not need to change the tests. Closing. Thank you @DimmestP.
