### 2022-05-17 fc6a56c
assert_that replaces some stopifnot calls
