### 2022-05-17 f72563b
cleanup create_colkey... as_factor
