### 2022-05-30 2487539
Added v0.5 to News in README.md
