### 2022-05-30 a61ec74
affiliations, grant numbers, rOpenSci, in paper.md
