### 2022-05-30 #470 @ewallace
Dear @jooolia,

We have finished updating tidyqpcr in response to your comments:

- Added missing files to .Rbuildignore and changed references to licence, version and depends in DESCRIPTION https://github.com/ewallace/tidyqpcr/commit/fa29569fcee2769dcfb14526a843aadafdf6f1ba
- made packages explicit in tests https://github.com/ewallace/tidyqpcr/commit/d31933b59542593ddf372b0a12656bc14c95041c
- getting started moved in readme https://github.com/ewallace/tidyqpcr/commit/c05f925b1b8f0b4f2295cea165f646cb9d490a5d
- Added top-level help-file https://github.com/ewallace/tidyqpcr/commit/98f930b7143c2dec68ba2e52e71706103f39b501
- Adds read_qpcr_data.R tests https://github.com/ewallace/tidyqpcr/commit/f5816f1858607b1ec512edff5c56563c4fb52508
- Removes {lifecycle} from description https://github.com/ewallace/tidyqpcr/commit/da45200631965af2b402a1b22bcdf398460d1007
- Address spelling errors https://github.com/ewallace/tidyqpcr/commit/cfe2e7c5c509853b1d10e9788f94a4a20bcec3d6 and 8f354b5743849f3c6f47068b9e1c79cbb91d3cdb
- calculate_dydx_1 to calculate_dydx https://github.com/ewallace/tidyqpcr/commit/45649a114b1878951e5fea036b3d9d1b34c5d359
- changes label_plate_rowcol warnings to notes https://github.com/ewallace/tidyqpcr/commit/ba1eb67493c594f249bdf6c4d8306943f1c7103b
- display_plate_qpcr() to enable display_plate() to be more general https://github.com/ewallace/tidyqpcr/commit/bd482f225bd77a58ab2bb5fd73bb7a8a4a5ec3fd
- improved plot_helpers.R styling https://github.com/ewallace/tidyqpcr/commit/55f4ca5104b2b64394d685a938a9f9f5f32384e6
- Adds basic usage to the Readme https://github.com/ewallace/tidyqpcr/commit/25bed2de902f5a095354699a2253988b76667bf5
- Updates pkgdown website https://github.com/ewallace/tidyqpcr/commit/b88ed9f66e399c56dcc2771f4fb955f923fc7841

We created a new tag [tidyqpcr v0.5](https://github.com/ewallace/tidyqpcr/releases/tag/v0.5), for the latest version.

We hope that we have fully addressed all your points, and look forward to your feedback.

Yours,
Sam and Edward
