### 2022-06-01 #470 @jooolia
Dear @ewallace,
Thanks for all your work and comments. I am having a look, but have been slowed down by devtools:check() exiting with status 1. I think it is something on my end, but I would like to understand what is happening. I will get back to you again on Friday. thanks, Julia
