### 2022-06-03 #470 @jooolia
Hi @ewallace ,

Thanks for all the work and for the comment pointing us to your updates. I haven't figured out what was causing me to have a exit status 1 with running "check" from the build window. I will ask about this on the rOpenSci slack. R CMD check and devtools::check() from the console are fine.

Regarding the package I have a few more little comments:

What is the story on the plot_helpers functions? Will these be further developed or incorporated in the future?

# URLs:

Not urgent, but can be fixed if desired.

## In platesetup_vignette.Rmd

- there is a missing "s" in the Issues address  - should be https://github.com/ewallace/tidyqpcr/issues (not issue)
- The cheatsheet https://rstudio.com/wp-content/uploads/2015/02/data-wrangling-cheatsheet.pdf has been moved to https://www.rstudio.com/wp-content/uploads/2015/02/data-wrangling-cheatsheet.pdf

## In the README.md

- https://www.force11.org/group/fairgroup/fairprinciples has been moved to https://force11.org/info/the-fair-data-principles/

# Tests

- In test-calculate_efficiency.R, I think that wrapping the function call on line 29 in suppressWarnings() would help with the interpretation of the tests here. Otherwise you get two warnings, but the warnings are not related to the test results and this could be confusing to the tester.

- label_plate_rowcol() - is not actually tested with a plate with well_col or well_row that are not factors. Do you think it would be important to test this? (This is not facetious or sarcastic comment, just genuinely thinking about it)

# Code
## dplyr joins:

In most uses of the "joins" in dplyr the "by" is specified according to good practice, however there are a few cases where they are not:

- debaseline() line 47 in amp_melt_curve_functions.R
- less urgent, but if there is time it would be good to fix this in the vignettes where the "by" is generally not specified.

## examples:
- (optional) The label_plate_rowcol() test has cases that don't throw warnings, maybe at some point it  could be helpful to add to the examples because to clarify how users could construct their plates.
