### 2022-06-03 ewallace/tidyqpcr#179 @ewallace
Fix URLs
From
https://github.com/ropensci/software-review/issues/470#issuecomment-1145867283

>
> # URLs:
>
> Not urgent, but can be fixed if desired.
>
> ## In platesetup_vignette.Rmd
>
> - there is a missing "s" in the Issues address  - should be https://github.com/ewallace/tidyqpcr/issues (not issue)
> - The cheatsheet https://rstudio.com/wp-content/uploads/2015/02/data-wrangling-cheatsheet.pdf has been moved to https://www.rstudio.com/wp-content/uploads/2015/02/data-wrangling-cheatsheet.pdf
>
> ## In the README.md
>
> - https://www.force11.org/group/fairgroup/fairprinciples has been moved to https://force11.org/info/the-fair-data-principles/
>
