### 2022-06-03 ewallace/tidyqpcr#181 @ewallace
Specify by argument in joins
From https://github.com/ropensci/software-review/issues/470#issuecomment-1145867283

> ## dplyr joins:
>
> In most uses of the "joins" in dplyr the "by" is specified according to good practice, however there are a few cases where they are not:
>
> - debaseline() line 47 in amp_melt_curve_functions.R
> - less urgent, but if there is time it would be good to fix this in the vignettes where the "by" is generally not specified.
