+++ R/amp_melt_curve_functions.R
@@ -57,7 +57,7 @@ debaseline <- function(plateamp, maxcycle = 10) {
                       .data$cycle <= maxcycle) %>%
         dplyr::summarize(fluor_base = stats::median(.data$fluor_raw))
     plateamp %>%
-        dplyr::left_join(baseline) %>%
+        dplyr::left_join(baseline, by = "well") %>%
         dplyr::mutate(fluor_signal = .data$fluor_raw - .data$fluor_base)
 }
 
