+++ vignettes/multifactor_vignette.Rmd
@@ -126,7 +126,7 @@ file_path_cq_plate1 <- system.file("extdata",
 
 plate1 <- file_path_cq_plate1 %>%
   read_lightcycler_1colour_cq() %>%
-  left_join(plateplan) %>%
+  left_join(plateplan, by = "well") %>%
   mutate(biol_rep = "1", plate = "1")
 
 file_path_cq_plate2 <- system.file("extdata",
@@ -135,7 +135,7 @@ file_path_cq_plate2 <- system.file("extdata",
 
 plate2 <- file_path_cq_plate2 %>%
   read_lightcycler_1colour_cq() %>%
-  left_join(plateplan) %>%
+  left_join(plateplan, by = "well") %>%
   mutate(biol_rep = "2", plate = "2")
 
 # combine data from both plates into a single data frame
@@ -276,7 +276,7 @@ file_path_raw_plate1 <- system.file("extdata/Edward_qPCR_TxnInhibitors_HS_2018-0
 plate1curve <- file_path_raw_plate1 %>%
     read_lightcycler_1colour_raw() %>%
     debaseline() %>%
-    left_join(plateplan) %>%
+    left_join(plateplan, by = "well") %>%
     mutate(biol_rep = 1, plate = 1)
 
 file_path_raw_plate2 <- system.file("extdata/Edward_qPCR_TxnInhibitors_HS_2018-06-15_plate2.txt.gz",
@@ -285,7 +285,7 @@ file_path_raw_plate2 <- system.file("extdata/Edward_qPCR_TxnInhibitors_HS_2018-0
 plate2curve <- file_path_raw_plate2 %>%
     read_lightcycler_1colour_raw() %>%
     debaseline() %>%
-    left_join(plateplan) %>%
+    left_join(plateplan, by = "well") %>%
     mutate(biol_rep = 2, plate = 2)
 
 platesamp <- bind_rows(plate1curve, plate2curve) %>%
