### 2022-06-03 198120c
delete plot_helpers, use scales::label_number
