### 2022-06-03 e32d9f1
test display_plate variants separately
