### 2022-06-03 21db8dc
clarify tests for plate functions
- give different test variables different names to reduce risks
- test factor coercion separately for rows and columns
- format code to avoid overly long lines
