+++ tests/testthat/test-plate_functions.R
@@ -87,7 +87,8 @@ test_that("functions for creating a 384 well, primer calibration plate work with
     
     calculated_plate_schematic <- display_plate_qpcr(calculated_labeled_384_plate)
     
-    calculated_plate_value <- display_plate_value(calculated_labeled_384_plate, value = "dilution")
+    calculated_plate_value <- display_plate_value(calculated_labeled_384_plate, 
+                                                  value = "dilution")
 
     expect_equal(calculated_colkey,
                  simulated_colkey)
@@ -183,69 +184,77 @@ test_that("functions for creating a 384 well, primer calibration plate work with
 
 ### Create dataset of expected function outputs ###
 
-simulated_colkey <- tibble(well_col = 1:24,
-                      biol_rep = factor(rep(c("A", "B",
-                                              "C", "D"),
-                                            each = 6),
-                                        levels = c("A", "B",
-                                                   "C", "D")),
-                      sample_id = rep(c("Sample_1", "Sample_2",
-                                        "Sample_3", "Sample_4",
-                                        "Sample_5", "Sample_6"),
-                                          times = 4))
-simulated_rowkey <- tibble(well_row = factor(LETTERS[1:16]),
-                      prep_type = factor(rep(c("+RT", "+RT",
-                                               "+RT", "-RT"),
-                                             each = 4),
-                                         levels = c("+RT", "-RT")),
-                      tech_rep = factor(rep(c(1, 2, 3, 1),
-                                            each = 4),
-                                        levels = c(1:3)),
-                      target_id = rep(c("Target_1", "Target_2",
-                                        "Target_3", "Target_4"),
-                                      times = 4))
+# colkey has well_col a number not a factor, testing coercion
+simulated_colkey_6_in_24 <- 
+    tibble(well_col = 1:24,
+           biol_rep = factor(rep(c("A", "B",
+                                   "C", "D"),
+                                 each = 6),
+                             levels = c("A", "B",
+                                        "C", "D")),
+           sample_id = rep(c("Sample_1", "Sample_2",
+                             "Sample_3", "Sample_4",
+                             "Sample_5", "Sample_6"),
+                           times = 4))
 
-simulated_label_plate <- tibble(well = paste0(rep(LETTERS[1:16], each = 24),
-                                              rep(1:24, times = 16)),
-                           well_row = factor(rep(LETTERS[1:16], each = 24)),
-                           well_col = factor(rep(1:24, times = 16)),
-                           biol_rep = factor(rep(rep(c("A", "B",
-                                                       "C", "D"),
-                                                     each = 6),
-                                                 times = 16),
-                                             levels = c("A", "B",
-                                                        "C", "D")),
-                           sample_id = rep(c("Sample_1", "Sample_2",
-                                             "Sample_3", "Sample_4",
-                                             "Sample_5", "Sample_6"),
-                                           times = 64),
-                           prep_type = factor(rep(c("+RT", "+RT",
-                                                    "+RT", "-RT"),
-                                                      each = 96),
-                                                  levels = c("+RT", "-RT")),
-                           tech_rep = factor(c(rep(1:3, each = 96),
-                                               rep(1, times = 96))),
-                           target_id = rep(rep(c("Target_1", "Target_2",
-                                                 "Target_3", "Target_4"),
-                                               each = 24),
-                                           times = 4))
+
+simulated_rowkey_4_in_16 <- 
[…]
