### 2022-06-03 ewallace/tidyqpcr#180 @ewallace
I addressed this in #185.

Adding examples for `label_plate_rowcol()` including factor coercion gives the user the chance to understand that works, and illustrates the points in the documentation.
Those examples are also run during package checking.

We have 3 tests of `label_plate_rowcol()`. Edited so one coerces neither, one coerces only rows, one coerces only columns.
