### 2022-06-08 #470 @ewallace
Hi @jooolia,

Thank you for this thorough check, it has helped a lot and given us more opportunity for improvements. @DimmestP and I have addressed these:

- plot_helpers. Given the minimal use of these functions and that their capabilities are now in the scales package, we removed these in pull request https://github.com/ewallace/tidyqpcr/pull/183
- URLs. Fixed in https://github.com/ewallace/tidyqpcr/pull/184
- Tests. Addressed in https://github.com/ewallace/tidyqpcr/pull/185
- dplyr joins. Addressed in https://github.com/ewallace/tidyqpcr/pull/182

We hope that's everything for the review process? Let us know.

Also we are open to suggestions as to whether it is better to share via CRAN or via Bioconductor. I really don't know.
