### 2022-06-10 #470 @ropensci-review-bot
Approved! Thanks @ewallace for submitting and @kelshmo for your reviews! :grin:
