### 2022-06-15 #470 @ewallace
- [ ] Transfer the repo to rOpenSci's "ropensci" GitHub organization under "Settings" in your repo.
- [ ] After transfer write a comment @ropensci-review-bot finalize transfer of <package-name> where <package-name> is the repo/package name. This will give you admin access back.
- [ ] If you already had a pkgdown website and are ok relying only on [rOpenSci central docs building and branding](https://devguide.ropensci.org/ci.html#even-more-ci-ropensci-docs),
  - [ ] deactivate the automatic deployment you might have set up
  - [ ] remove styling tweaks from your pkgdown config but keep that config file
  - [ ] replace the whole current pkgdown website with a [redirecting page] (https://devguide.ropensci.org/redirect.html)
  - [ ] replace your package docs URL with https://docs.ropensci.org/package_name
In addition, in your DESCRIPTION file, include the docs link in the URL field alongside the link to the GitHub repository, e.g.: URL: https://docs.ropensci.org/foobar, https://github.com/ropensci/foobar
- [ ] You can add this installation method to your package README install.packages("<package-name>", repos = "https://ropensci.r-universe.dev") thanks to [R-universe](https://ropensci.org/blog/2021/06/22/setup-runiverse/).
- [x] Submit to [JOSS](https://joss.theoj.org/), [tidyqpcr review page at JOSS](https://joss.theoj.org/papers/51996e6ce004532d31c0d83a6a806334)
- [ ] Fix [CRAN gotchas](https://devguide.ropensci.org/building.html#crangotchas)
